% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupar_datas.R
\name{agrupar_datas}
\alias{agrupar_datas}
\title{Subdivide as data em intervalos anuais, semestrais, trimestrais ou mensais.}
\usage{
agrupar_datas(
  data_inicial = NULL,
  data_final = NULL,
  corte = c("anual", "semestral", "trimestral", "mensal"),
  formato = "\%d/\%m/\%Y"
)
}
\arguments{
\item{data_inicial}{Data inicial dd/mm/aaaa}

\item{data_final}{Data final dd/mm/aaaa}

\item{corte}{Informar se anual, semestral, trimestral ou mensal.}

\item{formato}{Formato de retorno. Default para dd/mm/aaaa.}
}
\value{
tibble com duas colunas com data inicial e final
}
\description{
Subdivide as data em intervalos anuais, semestrais, trimestrais ou mensais.
}
