% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusta_assuntos.R
\name{ajusta_assuntos}
\alias{ajusta_assuntos}
\title{Ajustar os assuntos com base na tabela do TJSP}
\usage{
ajusta_assuntos(df, excluir_cartas = TRUE)
}
\arguments{
\item{df}{tibble obtido a partir da função organiza_dados_cpopg}

\item{excluir_cartas}{default para TRUE. Excluir cartas de ordem e pracatórias.}
}
\value{
mesmo df com assuntos ajustados e áreas acrescidas.
}
\description{
Ajustar os assuntos com base na tabela do TJSP
}
