% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar_assuntos.R
\name{ajustar_assuntos}
\alias{ajustar_assuntos}
\title{Ajustar os assuntos com base na tabela do TJSP}
\usage{
ajustar_assuntos(df)
}
\arguments{
\item{df}{tibble obtido a partir da função organiza_dados_cpopg}
}
\value{
mesmo df com assuntos ajustados e áreas acrescidas.
}
\description{
Ajustar os assuntos com base na tabela do TJSP
}
