% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assuntos.R
\docType{data}
\name{assuntos}
\alias{assuntos}
\title{assuntos}
\format{
dataframe com 3614 obs e três colunas:
\describe{
  \item{\code{area}}{character área conforme tabela do TJSP}
  \item{\code{assunto}}{character assunto com código}
  \item{\code{assunto_}}{character assunto sem código, sem acento, minúsculo}
}
}
\source{
\url{https://tjsp.jus.br/}
}
\usage{
assuntos
}
\description{
Assuntos do TJSP
}
\keyword{datasets}
