% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_acordaos.R
\name{baixar_acordao}
\alias{baixar_acordao}
\title{Baixa ac\\u00f3rd\\u00e3os do Tribunal de Justi\\u00e7a de S\\u00e3o Paulo}
\usage{
baixar_acordao(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{vetor com o N\\u00famero do processo com os separadores}

\item{diretorio}{Diret\\u00f3rio. Default para o corrente.}
}
\value{
tabela com metadados do pdf e baixa o pdf do ac\\u00f3rd\\u00e3o com o n\\u00famero como nome.
}
\description{
Baixa ac\\u00f3rd\\u00e3os do Tribunal de Justi\\u00e7a de S\\u00e3o Paulo
}
