% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_acordaos.R, R/baixar_acordaos2.R
\name{baixar_acordaos}
\alias{baixar_acordaos}
\title{Baixa acórdãos do Tribunal de Justiça de São Paulo a partir do andamento}
\usage{
baixar_acordaos(processos = NULL, diretorio = ".")

baixar_acordaos(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{vetor com o número do processo com ou sem os separadores}

\item{diretorio}{Diretório. Default para o corrente.}
}
\value{
tabela com metadados do pdf e pdf do acórdão baixado com data
   do acordao e número do processo.

tabela com metadados do pdf e pdf do acórdão baixado com data
   do acordao e número do processo.
}
\description{
Baixa acórdãos do Tribunal de Justiça de São Paulo a partir do andamento

Baixa acórdãos do Tribunal de Justiça de São Paulo a partir do andamento
}
