% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cjpg.r
\name{baixar_cjpg}
\alias{baixar_cjpg}
\title{Baixa julgados de primeiro grau de São Paulo}
\usage{
baixar_cjpg(livre = "", aspas = FALSE, processo = "", foro = "",
  vara = "", classe = "", assunto = "", magistrado = "",
  inicio = "", fim = "", diretorio = ".")
}
\arguments{
\item{livre}{Busca livre. Usar os mesmos critérios da página.}

\item{aspas}{Lógico. Coloque TRUE se quiser colocar aspas nas busca livre.}

\item{processo}{Número do processo unificado. Pode colocar com pontos ou sem pontos.}

\item{foro}{Número do fórum}

\item{vara}{Número da vara}

\item{classe}{Número da classe processual}

\item{assunto}{Número do assunto}

\item{magistrado}{Código do magistrado}

\item{inicio}{Data inicial no formato "dd/mm/aaa"}

\item{fim}{Data final no formato "dd/mm/aaaa" O lapso entre essas duas datas não deve passar de um ano.}

\item{diretorio}{Diretório onde serão armazenados os htmls.}
}
\value{
Baixa os htmls com metadados das decisãoes de primeiro grau.
}
\description{
Baixa julgados de primeiro grau de São Paulo
}
\examples{
\dontrun{
baixar_cjpg("homicídio simples")
}

}
