% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cjpg_processo.R
\name{baixar_cjpg_processo}
\alias{baixar_cjpg_processo}
\title{Baixa processo da consulta do cjpg}
\usage{
baixar_cjpg_processo(processo = NULL, diretorio = ".")
}
\arguments{
\item{processo}{número do processo}

\item{diretorio}{diretório}
}
\value{
html
}
\description{
Baixa processo da consulta do cjpg
}
