% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cposg.R
\name{baixar_cposg}
\alias{baixar_cposg}
\title{Baixa htmls dos processos de segunda instância do TJSP}
\usage{
baixar_cposg(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{vetor com os números dos processos, com ou sem os separadores.}

\item{diretorio}{Diretório onde devem ser armazenados os htmls}
}
\value{
Arquivos htmls no formato aaaa_mm_dd_OOOOOOODDAAAAJTRNNNN.html indicando a data
     do download seguida pelo número do processo.
}
\description{
Baixa htmls dos processos de segunda instância do TJSP
}
\examples{
\dontrun{
baixar_cposg("feminicídio")
}
}
