% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_municipios.R
\name{baixar_municipios}
\alias{baixar_municipios}
\title{Baixa municípios do interior de São Paulo com
    respectivos fóruns}
\usage{
baixar_municipios(numeros = 1:645, diretorio = ".")
}
\arguments{
\item{numeros}{Sequência numérica}

\item{diretorio}{Onde armazenar os htms}
}
\value{
htm
}
\description{
Baixa municípios do interior de São Paulo com
    respectivos fóruns
}
