% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_publicacoes_dje.R
\name{baixar_publicacoes_dje}
\alias{baixar_publicacoes_dje}
\title{Baixar todas as publicações do dje São Paulo.}
\usage{
baixar_publicacoes_dje(processo, diretorio = ".")
}
\arguments{
\item{processo}{número do processo com ou sem pontos e hífen.}

\item{diretorio}{default para o diretório atual.}
}
\value{
publicações em html
}
\description{
Baixar todas as publicações do dje São Paulo.
}
\details{
Esta função não é vetorizada.
}
