% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_serializar.R
\name{baixar_serializar}
\alias{baixar_serializar}
\title{Aplica uma função para baixar htmls e serializa os htmls em grupos de
    no máximo aproximadamente 300 Mb.}
\usage{
baixar_serializar(lista = NULL, dir = ".", subdir = NULL, funcao = NULL)
}
\arguments{
\item{lista}{lista de vetores de processos ou simplesmente um vetor de processos}

\item{dir}{diretorio onde serão criados os subdiretorios}

\item{subdir}{subdiretórios onde serão armazenados  grupo da lista.}

\item{funcao}{Função de download do html, ex. esaj::download_cpopg}
}
\value{
Arquivo em rds baixado
}
\description{
Aplica uma função para baixar htmls e serializa os htmls em grupos de
    no máximo aproximadamente 300 Mb.
}
