% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg.R
\name{cjpg}
\alias{cjpg}
\title{Consulta de Julgados de Primeiro Grau}
\usage{
cjpg(session, parms, path = "data-raw/cjpg", max_pag = 10,
  overwrite = FALSE, verbose = TRUE, p = 0.05)
}
\arguments{
\item{session}{sessão da CJPG, retornado pela função \code{\link{cjpg_session}}.}

\item{parms}{form com parâmetros de busca, retornado pela função \code{\link{cjpg_parms}}.}

\item{path}{caminho onde os arquivos HTML serão baixados. Tentará criar a pasta automaticamente.}

\item{max_pag}{número máximo de páginas a serem baixadas. Se \code{NA} ou \code{Inf}, baixará todas as páginas resultantes da pesquisa.}

\item{overwrite}{logical: sobrescrever arquivos salvos?}

\item{verbose}{logical: imprimir mensagens de acompanhamento do download?}

\item{p}{probabilidade de imprimir uma mensagem de acompanhamento se \code{verbose} for \code{TRUE}. Default 5\%.}
}
\description{
Baixa arquivos HTML correspondentes a uma busca na CJPG.
}
