% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg.R
\name{cjpg_npags}
\alias{cjpg_npags}
\title{Número de páginas}
\usage{
cjpg_npags(session, parms = NULL)
}
\arguments{
\item{session}{sessão da CJPG.}

\item{parms}{se for \code{NULL}, admite que já está na página de resultados da CJPG.}
}
\value{
objeto de classe \code{session}.
}
\description{
Calcula o número de páginas retornadas por uma consulta de julgados de primeiro grau.
}
\examples{
\dontrun{
library(tjsp)
s <- cjsg_session()
parms <- cjsg_parms(s, livre = 'acordam')
cjsg_npags(s, parms)
cjsg(s, parms, max_pag = 1, path = '.') # chamada internamente dentro de cjsg.
}

}
