% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg.R
\name{cjpg_parms}
\alias{cjpg_parms}
\title{Parâmetros de pesquisa}
\usage{
cjpg_parms(session, livre = "", data_inicial = NULL, data_final = NULL,
  secoes = "", classes = "", assuntos = "", magistrados = "",
  varas = "")
}
\arguments{
\item{session}{sessão da CJPG.}

\item{livre}{string com pesquisa livre por palavras-chave.}

\item{data_inicial}{data inicial da decisão em formado \%Y-\%m-\%d ou objeto Date.}

\item{data_final}{data final da decisão em formado \%Y-\%m-\%d ou objeto Date.}

\item{classes}{vetor com as classes que se deseja pesquisar. Obter o vetor de \code{\link{cjpg_classes}}.}

\item{assuntos}{vetor com os assuntos que se deseja pesquisar. Obter o vetor de \code{\link{cjpg_assuntos}}.}

\item{magistrados}{vetor com os magistrados que se deseja pesquisar. NOT IMPLEMENTED YET.}

\item{varas}{vetor com as varas que se deseja pesquisar. Obter o vetor de \code{\link{cjpg_varas}}.}
}
\value{
objeto de classe \code{form}.
}
\description{
Inclui parâmetros de pesquisa da CJPG.
}
\details{
O intervalo de datas, caso seja incluído, precisa ser de no máximo um ano.
}
\examples{
\dontrun{
library(tjsp)
s <- cjpg_session()
parms <- cjsg_parms(s, livre = 'sentenca', data_inicial = '2015-01-01', data_final = '2015-05-01')
parms
cjpg(s, parms, max_pag = 1, path = '.')
}
}

