% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjpg_table.R
\name{cjpg_table}
\alias{cjpg_table}
\title{Extrai assuntos e classes}
\usage{
cjpg_table(tipo)
}
\arguments{
\item{tipo}{classe ou assunto}
}
\value{
tibble com classe ou assunto
}
\description{
Extrai assuntos e classes
}
\examples{
\dontrun{
df <- cjpg_table(tipo="classe")
}
}
