% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsg.R
\name{cjsg_pag}
\alias{cjsg_pag}
\title{Baixa uma página.}
\usage{
cjsg_pag(pag, path, ow, s)
}
\arguments{
\item{pag}{número da página a ser baixada.}

\item{path}{pasta em que o arquivo será salvo.}

\item{ow}{logical, sobrescrever arquivo?}

\item{s}{sessao da CJSG.}
}
\description{
Baixa uma página a partir de um resultado da CJSG. Não deve ser usado diretamente.
}

