% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects-2inst.R
\name{cjsg_tabs}
\alias{cjsg_tabs}
\title{Baixa tabela de itens da CJPG.}
\usage{
cjsg_tabs(tipo = c("classes", "assuntos"))
}
\arguments{
\item{tipo}{classe, assunto ou varas}
}
\value{
\code{data.frame} com titulos e cods das folhas e titulos e cods dos pais (índice 0 é a raiz e quanto menor o índice, mais próximo da raiz)
}
\description{
Baixa uma tabela com os elementos de classe, assunto ou varas, para ajudar a baixar usando a função CJPG.
}
