% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_recurso2.R
\name{classificar_recurso2}
\alias{classificar_recurso2}
\title{Classifica recursos}
\usage{
classificar_recurso2(x)
}
\arguments{
\item{x}{vetor de julgados}
}
\value{
decisão
}
\description{
Classifica recursos
}
\examples{
\dontrun{
x <- classificar_recurso2(x = julgado)
}
}
