% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_sentencas.R
\name{classificar_sentenca}
\alias{classificar_sentenca}
\title{Classifica decisões de primeiro grau}
\usage{
classificar_sentenca(x, sentenca, decisao)
}
\arguments{
\item{x}{data.frame cjpg}

\item{sentenca}{nome da coluna, sem aspas,  com as sentenças (julgados)}

\item{decisao}{nome da coluna, entre aspas, a ser criada.}
}
\value{
cria uma nova coluna com a classificação das decisões
}
\description{
Classifica decisões de primeiro grau
}
\examples{
\dontrun{
df <- classificar_sentenca(df,julgado,"decisao")
}
}
