% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificador_sentenca2.R
\name{classificar_sentenca2}
\alias{classificar_sentenca2}
\title{Classifica decisões de primeiro grau}
\usage{
classificar_sentenca2(sentenca)
}
\arguments{
\item{sentenca}{vetor de sentenças}
}
\value{
decisão
}
\description{
Classifica decisões de primeiro grau
}
\examples{
\dontrun{
x <- classificar_sentenca2(sentenca)
}
}
