% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_writ.R
\name{classificar_writ}
\alias{classificar_writ}
\title{Classificar writs(HC, MS, revisão criminal)}
\usage{
classificar_writ(texto)
}
\arguments{
\item{texto}{Texto da decisão}
}
\value{
classificação
}
\description{
Classificar writs(HC, MS, revisão criminal)
}
