% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_sequencial.R
\name{cnj_sequencial}
\alias{cnj_sequencial}
\title{Creates cnj sequence numbers}
\usage{
cnj_sequencial(inicio, fim, ano, segmento, uf, distribuidor)
}
\arguments{
\item{inicio}{integer}

\item{fim}{integer}

\item{ano}{year}

\item{segmento}{integer}

\item{uf}{federative unity.}

\item{distribuidor}{code of the court.}
}
\value{
vector of lawsuit numbers
}
\description{
Creates cnj sequence numbers
}
\examples{
\dontrun{
stj<-cnj_sequencial(0,100,2018,3,00,0000)
}
}
