% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\docType{data}
\name{codigos}
\alias{codigos}
\title{Códigos}
\format{
A data frame with 388 rows and 5 variables:
\describe{
  \item{\code{segmento}}{character segmento do Judiciário conforme resolução 65 CNJ}
  \item{\code{uf}}{double Código do estado conforme resolução 65 do CNJ}
  \item{\code{comarca}}{character Nome da comarca}
  \item{\code{distribuidor}}{double Código do distribuidor conforme resolução 65 CNJ}
  \item{\code{descricao}}{character Descrição do distribuidor}
}
}
\source{
\url{https://tjsp.jus.br/}
}
\usage{
codigos
}
\description{
Códigos dos distribuidores do Tribunal de Justiça de São Paulo
}
\keyword{datasets}
