% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos2.R
\docType{data}
\name{codigos2}
\alias{codigos2}
\title{Códigos2}
\format{
A data frame with 443 rows and 2 variables:
\describe{
  \item{\code{distribuidor}}{double Código do distribuidor conforme resolução 65 CNJ}
  \item{\code{descricao}}{character Descrição do distribuidor}
}
}
\source{
\url{https://tjsp.jus.br/}
}
\usage{
codigos2
}
\description{
Códigos dos distribuidores do Tribunal de Justiça de São Paulo
}
\keyword{datasets}
