% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cposg_entrada.R
\name{cposg_entrada}
\alias{cposg_entrada}
\title{Extrai data de entrada/recebimento da a\u00e7\u00e3o ou recurso em segunda inst\u00e2ncia}
\usage{
cposg_entrada(processos = NULL)
}
\arguments{
\item{processos}{}
}
\value{
data de entada do recurso
}
\description{
Extrai data de entrada/recebimento da a\u00e7\u00e3o ou recurso em segunda inst\u00e2ncia
}
\details{
Esta fun\u00e7\u00e3o foi criada somente para resolver retirar a entrada de
    processos que entraram mais de uma vez. Ela pega somente a \u00faltima data.
    Por ora, n\u00e3o vejo necessidade de pegar todas as entradas, vez que a finalidade
    \u00e9 somente calcular o tempo do processo.
}
