% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cposg_mov.R
\name{cposg_mov}
\alias{cposg_mov}
\title{Extrai a movimenta\\u00e7\\u00e3o processual de segunda inst\\u00e2ncia}
\usage{
cposg_mov(processos = NULL)
}
\arguments{
\item{processos}{n\\u00famero do processo de segunda inst\\u00e2ncia no formato NNNNNNN.DD.AAAA.J.TR.OOOO}
}
\value{
data frame com a movimenta\u00e7\u00e3o processual.
}
\description{
Extrai a movimenta\\u00e7\\u00e3o processual de segunda inst\\u00e2ncia
}
\examples{

movimentacao<-cposg_mov("1049363-11.2015.8.26.0002")
}
