% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decifrar.R
\name{decifrar}
\alias{decifrar}
\title{Decifra o audio}
\usage{
decifrar(arq_mp3)
}
\arguments{
\item{arq_mp3}{arquivo que contém o audio.}
}
\value{
character vector de tamanho 1 com as letras preditas.
}
\description{
Le um arquivo de audio e tenta decifrar quais sao as letras.
}
