% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_eproc_jurisprudencia.R
\name{eproc_montar_form}
\alias{eproc_montar_form}
\title{Worker interno para download de jurisprudência eproc}
\usage{
eproc_montar_form(body)
}
\arguments{
\item{base_url}{URL base do sistema eproc (1g ou 2g).}
}
\value{
Salva arquivos HTML no diretório especificado.
}
\description{
Worker interno para download de jurisprudência eproc
}
