% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_eproc_inteiro_teor.R
\name{eproc_parsear_artigo}
\alias{eproc_parsear_artigo}
\title{Parseia um artigo do inteiro teor do eproc}
\usage{
eproc_parsear_artigo(art, id_jurisprudencia)
}
\arguments{
\item{art}{Nó xml2 de um artigo.}

\item{id_jurisprudencia}{ID do arquivo de origem.}
}
\value{
Tibble de uma linha.
}
\description{
Parseia um artigo do inteiro teor do eproc
}
