% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_eproc_jurisprudencia.R
\name{eproc_parsear_item}
\alias{eproc_parsear_item}
\title{Parseia um item de resultado do eproc}
\usage{
eproc_parsear_item(item, pagina)
}
\arguments{
\item{item}{Nó xml2 de um resultado individual.}

\item{pagina}{Número da página de origem.}
}
\value{
Tibble de uma linha.
}
\description{
Parseia um item de resultado do eproc
}
