% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excluir_classes_cpopg.R
\docType{data}
\name{excluir_classes_cpopg}
\alias{excluir_classes_cpopg}
\title{excluir_classes_cpopg}
\format{
vector:
\describe{
  \item{\code{excluir_classes_cpopg}}{character with suggestions classes to remove}

}
}
\source{
\url{https://tjsp.jus.br/}
}
\usage{
excluir_classes_cpopg
}
\description{
Excluir classes
}
\keyword{datasets}
