% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtra.R
\name{filtra}
\alias{filtra}
\title{Function filtra}
\usage{
filtra(df, column, string)
}
\arguments{
\item{column}{column name where to search for string}

\item{string}{string to be found in the mentioned column}

\item{x}{data.frame}
}
\description{
This functions uses regex to filter a data frame based on a column's content
}
\examples{
filtra(df,"column","string")
}
\keyword{regex,filter}
