% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrar.R
\name{filtrar}
\alias{filtrar}
\title{Filtra tabela (dataframe) com base numa expressão regular.}
\usage{
filtrar(x, coluna, expressao)
}
\arguments{
\item{x}{tibble ou data.frame}

\item{coluna}{coluna}

\item{expressao}{expressão regular}
}
\description{
Filtra tabela (dataframe) com base numa expressão regular.
}
\details{
Esta função dispensa o uso de aspas para a coluna(NSE). Ela equivale a dplyr::filter(df,str_detect(coluna,expressao))
}
\examples{
df<-data.frame(a = c("furto", "roubo","tráfico"),
   b = c("roubo", "homicidio culposo","lesão corporal"),
   c = c("roubo", "ameaça","tráfico"))
   filtrar(df,c,"roubo")
}
\keyword{regex,filter}
