% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrar_cjsg.R
\name{filtrar_cjsg}
\alias{filtrar_cjsg}
\title{Filtra base cjsg por assunto e classe.}
\usage{
filtrar_cjsg(base, assunto, classe, negate_assunto = FALSE)
}
\arguments{
\item{base}{base parseada dos cjsg htmls}

\item{assunto}{regex do assunto a ser filtrado}

\item{classe}{regex da classe a ser filtrada}

\item{negate_assunto}{coloque TRUE se quiser excluir os assuntos especificados.}
}
\value{
Mesma base filtrada.
}
\description{
Filtra base cjsg por assunto e classe.
}
\details{
Geralmente as classes são limitadas e as descrições mais claras que
    os assuntos. Por isso optei pela negação somente do assunto.
}
