% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cjsg.R
\name{formatar_arquivo}
\alias{formatar_arquivo}
\title{Função para criar o nome do arquivo}
\usage{
formatar_arquivo(inicio, fim, inicio_pb, fim_pb, pagina, diretorio)
}
\arguments{
\item{inicio}{data inicial julgamento}

\item{fim}{Data final julgamento}

\item{inicio_pb}{data inicial registro/publicação}

\item{fim_pb}{data final registr/publicacao}

\item{pagina}{página}

\item{diretorio}{diretorio}
}
\value{
Arquivo
}
\description{
Função para criar o nome do arquivo
}
