% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_search.R
\name{fuzzy_search}
\alias{fuzzy_search}
\title{Procura a palavra ou frase do segundo vetor que melhor
   se aproxima do primeiro. Particularmente útil para
   comparar nomes de municípios.}
\usage{
fuzzy_search(x = NULL, y = NULL)
}
\arguments{
\item{x}{Vetor de strings supostamente erradas}

\item{y}{Vetor de strings suportamente corretas}
}
\value{
vetor com as strings de y próximos
    de x.
}
\description{
Procura a palavra ou frase do segundo vetor que melhor
   se aproxima do primeiro. Particularmente útil para
   comparar nomes de municípios.
}
\examples{
wrong_names<-c("Mogi das Cruses","Sao Paulo","CANTA GALLO")
correct_names<-c("Canta Galo","Mogi das Cruzes","São Paulo")
fuzzy_search(x=wrong_names,y=correct_names)
}
