% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdecisao.R
\name{ggdecisao}
\alias{ggdecisao}
\title{Plota distribuição das decisões conforme a câmara e o recorrente}
\usage{
ggdecisao(
  data,
  title = "",
  x = NULL,
  facet = NULL,
  fill = NULL,
  ordered = TRUE,
  x_label = "Câmara",
  y_label = "Número de decisões",
  legend = "Decisão",
  caption = "Fonte: TJSP"
)
}
\arguments{
\item{data}{dataframe}

\item{title}{Título do gráfico}

\item{x}{orgao julgador, geralmente a câmara}

\item{facet}{facetas para facet_grid, geralmente o recorrente}

\item{fill}{preenchimento, geralmente a decisão}

\item{ordered}{se verdadeiro irá extrair somente os números e as siglas
e ordená-la adequadaemente.}

\item{x_label}{etiqueta para órgão julgador, geralmente "câmara"}

\item{y_label}{etiqueta para número de casos, geralmente "número de decisões"}

\item{legend}{título da legenda}

\item{caption}{indicação da fonte}
}
\value{
ggplot
}
\description{
Plota distribuição das decisões conforme a câmara e o recorrente
}
