% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteiroTJSP.R
\name{inteiroTJSP}
\alias{inteiroTJSP}
\title{This function downloads complete decisions from São Paulo state High Court.}
\usage{
inteiroTJSP(url = NULL, cdacordao = NULL)
}
\arguments{
\item{url}{a vector of url strings to access the decisions.}

\item{cdacordao}{a vector of decisions ids/codes to give as names to the downloading files.}
}
\value{
pdf of the decision downloaded to the working directory.
}
\description{
This function downloads complete decisions from São Paulo state High Court.
}
\keyword{court}
\keyword{decisions}
\keyword{tjsp}
