% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_combinar_docs.R
\name{is_pdf}
\alias{is_pdf}
\title{Verifica se um arquivo é pdf (vetorizado)}
\usage{
is_pdf(x)
}
\arguments{
\item{x}{Vetor de arquivos}
}
\value{
Valor lógico
}
\description{
Verifica se um arquivo é pdf (vetorizado)
}
