% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_combinar_docs.R
\name{is_pdf1}
\alias{is_pdf1}
\title{Verifica se um arquivo é pdf (não vetorizado)}
\usage{
is_pdf1(arquivo)
}
\arguments{
\item{arquivo}{Suposto pdf a ser verificado}
}
\value{
Valor lógico
}
\description{
Verifica se um arquivo é pdf (não vetorizado)
}
