% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapso.R
\name{lapso}
\alias{lapso}
\title{Retorna o lapso temporal na unidade especificada entre duas datas.}
\usage{
lapso(data_inicial = NULL, data_final = NULL, unidade = "mes")
}
\arguments{
\item{data_inicial}{incluir a data no formato Date.}

\item{data_final}{incluir a data no formato Date.}

\item{unidade}{uma das opções: ano, mes, semana, dia.}
}
\value{
double com o lapso temporal
}
\description{
Retorna o lapso temporal na unidade especificada entre duas datas.
}
\details{
Esta função é um mero wrapper das funções interval e time_length
    do pacote lubridate. Ela considera as variações de ano, meses e dias.
}
\examples{
lapso(data_inicial = as.Date("2018-01-01"), data_final = as.Date("2018-12-31"), unidade = "mes")
}
