% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong_comarca.R
\name{latlong_comarca}
\alias{latlong_comarca}
\title{Encontra latlong das comarcas}
\usage{
latlong_comarca(df)
}
\arguments{
\item{df}{tibble com nomes das comarcas.}
}
\value{
mesma tibble com latlongs
}
\description{
Encontra latlong das comarcas
}
