% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decifrar.R
\name{ler}
\alias{ler}
\title{Lê e processa um arquivo mp3.}
\usage{
ler(arq_mp3)
}
\arguments{
\item{arq_mp3}{arquivo mp3 a ser lido.}
}
\value{
um vetor de inteiros com a soma dos valores absolutos das frequencias
de cada trecho identificado. Esses numeros podem ser considerados como
uma fingerprint da letra, pois nenhuma outra letra possui a mesma soma
de frequencias.
}
\description{
Le um arquivo mp3, com o pacote \code{tuneR}.
}
