% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_acordaos.R
\name{ler_acordaos}
\alias{ler_acordaos}
\title{Lê pdfs dos acórdãos, remove assinatura e eventualmente combina}
\usage{
ler_acordaos(diretorio = ".", remover_assinatura = FALSE, combinar = FALSE)
}
\arguments{
\item{diretorio}{diretório onde se encontram os pdfs}

\item{remover_assinatura}{default é FALSE}

\item{combinar}{default juntar todas as páginas em um único vetor.}
}
\value{
Lista com vetores das páginas dos pdfs
}
\description{
Lê pdfs dos acórdãos, remove assinatura e eventualmente combina
}
