% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cjpg.R
\name{ler_cjpg}
\alias{ler_cjpg}
\title{Lê decisões de primeira instância do TJSP baixadas por meio da função baixar_cjpg.}
\usage{
ler_cjpg(fonte = ".")
}
\arguments{
\item{fonte}{objeto ou diretório onde se encontram as páginas.}
}
\value{
uma tibble com nove colunas: processo, classe, assunto, magistrado,comarca, foro, vara,
    disponibilizacao e julgado (texto da decisão).
}
\description{
Lê decisões de primeira instância do TJSP baixadas por meio da função baixar_cjpg.
}
\examples{
\dontrun{
ler_cjpg()
}
}
