% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cjsg.R
\name{ler_cjsg}
\alias{ler_cjsg}
\title{Ler metadados das decisões}
\usage{
ler_cjsg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Caminhos para os htmls}

\item{diretorio}{Diretório onde se encontram os htmls,
se arquivos for NULL.}
}
\value{
tabela com metadados jurisprudenciais.
}
\description{
Ler metadados das decisões
}
\examples{
\dontrun{
cjsg <- ler_cjsg()
}

}
