% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados.R
\name{ler_dados}
\alias{ler_dados}
\title{Aplica parser para extrair informa\u00e7\u00f5es sobre as dados do
    processo.}
\usage{
ler_dados(path = ".")
}
\arguments{
\item{path}{diret\u00f3rio onde se encontram os htmls baixados.}
}
\value{
tabela com dados do processo
}
\description{
Aplica parser para extrair informa\u00e7\u00f5es sobre as dados do
    processo.
}
\examples{
dados<-ler_dados()

}
