% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cpopg_serializados.R
\name{ler_dados_cpopg_serializados}
\alias{ler_dados_cpopg_serializados}
\title{Lê dados do cpopg já serializados}
\usage{
ler_dados_cpopg_serializados(arquivo, wide = FALSE)
}
\arguments{
\item{arquivo}{arquivo rds serializado}

\item{wide}{lógico.long, default, ou wide.}
}
\value{
tibble em long ou wide format.
}
\description{
Lê dados do cpopg já serializados
}
