% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cposg.R
\name{ler_dados_cposg}
\alias{ler_dados_cposg}
\title{Lê metadados dos processos de segunda instância}
\usage{
ler_dados_cposg(fonte = ".")
}
\arguments{
\item{fonte}{objeto no global env ou diretório onde se encontram os htmls.}
}
\value{
tabela com dados do processo
}
\description{
Lê metadados dos processos de segunda instância
}
\examples{
\dontrun{
dados <- ler_dados_cposg()
}
}
