% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cposg2.R
\name{ler_dados_cposg2}
\alias{ler_dados_cposg2}
\title{Lê dados cposg, em uma das interfaces.}
\usage{
ler_dados_cposg2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório, quando arquivos não informados.}
}
\value{
dataframe
}
\description{
Lê dados cposg, em uma das interfaces.
}
