% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_decisao.R
\name{ler_decisoes}
\alias{ler_decisoes}
\title{Ler decisoes}
\usage{
ler_decisoes(path = ".")
}
\arguments{
\item{path}{Diretório onde se encontram os htmls baixados}
}
\value{
tibble com as os numéros dos processos e respectivas decisões
}
\description{
Ler decisoes
}
