% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_decisoes_cposg.R
\name{ler_decisoes_cposg}
\alias{ler_decisoes_cposg}
\title{Lê o dispositivo das decisões de segunda instância a partir dos htmls}
\usage{
ler_decisoes_cposg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Informar arquivos ou diretório.}

\item{diretorio}{Diretório onde se encontram os htmls baixados}
}
\value{
tibble com as os numéros dos processos e respectivos dispositivos
}
\description{
Lê o dispositivo das decisões de segunda instância a partir dos htmls
}
\examples{
\dontrun{
decisoes <- ler_decisoes_cposg()
}
}
