% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_decisoes_monocraticas.R
\name{ler_decisoes_monocraticas}
\alias{ler_decisoes_monocraticas}
\title{Lê conteúdo das decisões monocráticas}
\usage{
ler_decisoes_monocraticas(path = ".")
}
\arguments{
\item{path}{Diretório onde se encontram os htmls}
}
\value{
tibble com os htmls.
}
\description{
Lê conteúdo das decisões monocráticas
}
\examples{

ler_decisoes_monocraticas()

}
