% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_despachos.R
\name{ler_despacho}
\alias{ler_despacho}
\title{Extrai despachos judiciais, que podem incluir decisões monocráticas, da movimentação.}
\usage{
ler_despacho(fonte = ".")
}
\arguments{
\item{fonte}{objeto ou diretório onde se encontram os htmls.}
}
\value{
tibble com o número do processo, o despacho e a data do despacho.
}
\description{
Extrai despachos judiciais, que podem incluir decisões monocráticas, da movimentação.
}
\examples{
\dontrun{
ler_despachos()
}

}
