% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao.R
\name{ler_movimentacao}
\alias{ler_movimentacao}
\title{Ler o andamento dos processos}
\usage{
ler_movimentacao(path = ".")
}
\arguments{
\item{path}{Diretório onde se encontram os htmls dos processos.}
}
\value{
tibble com os processos e andamentos.
}
\description{
Ler o andamento dos processos
}
