% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao_cposg.R
\name{ler_movimentacao_cposg}
\alias{ler_movimentacao_cposg}
\title{Extrai a movimentação processual de segunda instância}
\usage{
ler_movimentacao_cposg(processos = NULL)
}
\arguments{
\item{processos}{número do processo de segunda instância no formato NNNNNNN.DD.AAAA.J.TR.OOOO}
}
\value{
data frame com a movimentação processual.
}
\description{
Extrai a movimentação processual de segunda instância
}
