% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_municipios.R
\name{ler_municipios}
\alias{ler_municipios}
\title{Lê municípios e respectivos fóruns baixados com baixar_municipios}
\usage{
ler_municipios(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Alternativamente informar diretório}
}
\value{
tibble
}
\description{
Lê municípios e respectivos fóruns baixados com baixar_municipios
}
