% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_partes.R
\name{ler_partes}
\alias{ler_partes}
\title{Aplica parser para extrair informações sobre as partes do
    processo.}
\usage{
ler_partes(diretorio = ".", arquivos = NULL)
}
\arguments{
\item{diretorio}{Objeto ou diretório onde se encontram os htmls baixados.}

\item{arquivos}{O diretório é ignorado se você fornecer o vetor
de arquivos}
}
\value{
tabela com informações das partes.
}
\description{
Aplica parser para extrair informações sobre as partes do
    processo.
}
