% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpar.R
\name{limpar}
\alias{limpar}
\title{Função limpar}
\usage{
limpar(texto)
}
\description{
Esta função limpa o texto em português: retira acentos, retira pontuação, coloca em caixa baixa
}
\examples{
limpa(texto)
}
\keyword{limpar,}
\keyword{stop}
\keyword{words,}
