% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects-2inst.R
\name{list_secoes_2inst}
\alias{list_secoes_2inst}
\title{Seções de segunda instância.}
\usage{
list_secoes_2inst()
}
\value{
\code{tibble} com as colunas \code{pai}, \code{secao} e \code{cod}. Usualmente \code{cod} será usado na função \code{\link{cjsg_parms}}
}
\description{
Baixa lista das seções de segunda instância.
}
