% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_base64_doc.R
\name{mni_base64_doc}
\alias{mni_base64_doc}
\title{Converte documentos em base 64 para pdf.}
\usage{
mni_base64_doc(arquivos = NULL, dir_origem = ".", dir_destino = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{dir_origem}{Alternativamente informar origem}

\item{dir_destino}{Diretório destino.}
}
\value{
pdfs
}
\description{
Converte documentos em base 64 para pdf.
}
