% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_combinar_docs.R
\name{mni_combinar_docs}
\alias{mni_combinar_docs}
\title{Combina docs baixados com a partir de mni_base64_docs}
\usage{
mni_combinar_docs(
  arquivos = NULL,
  dir_origem = ".",
  dir_destino = NULL,
  nivel = c("processo", "doc")
)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{dir_origem}{Diretório de origem, se não
informar arquivos}

\item{dir_destino}{Diretório destino}

\item{nivel}{Juntar processo inteiro ou por documento?}
}
\value{
único pdf
}
\description{
Combina docs baixados com a partir de mni_base64_docs
}
