% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_consultar_processo.R
\name{mni_consultar_processo}
\alias{mni_consultar_processo}
\title{Consulta dados processuais pelo Modelo de interoperabilidade (MNI)}
\usage{
mni_consultar_processo(
  tribunal = "tjsp",
  usuario = NULL,
  senha = NULL,
  processo = "",
  cabecalho = TRUE,
  movimentos = FALSE,
  documentos = FALSE,
  diretorio = "."
)
}
\arguments{
\item{tribunal}{Padrão "tjsp"}

\item{usuario}{Informar usuario ou variável de ambiente
TJSPMNIUSUARIO}

\item{senha}{Informar senha ou variável de ambiente
TJSPMNISENHA}

\item{processo}{Processo (número único CNJ)}

\item{cabecalho}{TRUE para pegar cabeçalho}

\item{movimentos}{TRUE para pegar movimentacao}

\item{documentos}{TRUE para pegar documentos.}

\item{diretorio}{Diretório}
}
\value{
xmls
}
\description{
Consulta dados processuais pelo Modelo de interoperabilidade (MNI)
}
