% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_ler_dados_basicos.R
\name{mni_ler_dados_basicos}
\alias{mni_ler_dados_basicos}
\title{Lê dados básicos xml mni}
\usage{
mni_ler_dados_basicos(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de xmls}

\item{diretorio}{Alternativamente informar diretório.}
}
\value{
tibble
}
\description{
Lê dados básicos xml mni
}
