% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_ler_partes.R
\name{mni_ler_partes}
\alias{mni_ler_partes}
\title{Lê informações das partes do xml baixado com mni_consultar_processo}
\usage{
mni_ler_partes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos xml.}

\item{diretorio}{Alternativamente informar o diretório.}
}
\value{
Tibble
}
\description{
Lê informações das partes do xml baixado com mni_consultar_processo
}
