% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_ler_xml_as_json.R
\name{mni_ler_xml_as_json}
\alias{mni_ler_xml_as_json}
\title{Lê xmls baixados com mni_consultar_processo como jsons}
\usage{
mni_ler_xml_as_json(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos.}

\item{diretorio}{Alternativamente informar diretório.}
}
\value{
Tibble com coluna processo e coluna json(jq)
}
\description{
Lê xmls baixados com mni_consultar_processo como jsons
}
