% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_listar_documentos.R
\name{mni_listar_documentos}
\alias{mni_listar_documentos}
\title{Listar documentos com respectivos códigos}
\usage{
mni_listar_documentos(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Alternativamente informar diretório}
}
\value{
tibble
}
\description{
Listar documentos com respectivos códigos
}
