% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelar_decisao.R
\name{modelar_decisao}
\alias{modelar_decisao}
\title{Modelar decisão}
\usage{
modelar_decisao(..., data, y = NULL, metodo = "glm")
}
\arguments{
\item{...}{Fórmula conforme modelo}

\item{data}{base}

\item{y}{variável dependente}

\item{metodo}{c("glm","gbm","xgboost","rf","catboost")}
}
\value{
objeto caret
}
\description{
Modelar decisão
}
