% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios.R
\docType{data}
\name{municipios}
\alias{municipios}
\title{municipios}
\format{
Dataframe com 725 linhas e 3 variáveis:
\describe{
  \item{\code{codigo}}{código da url para obter o dado}
  \item{\code{municipio}}{Município}
  \item{\code{forum}}{Fórum}
   }
}
\source{
\url{https://www.emsampa.com.br/delpol/sp_001.htm}
}
\usage{
municipios
}
\description{
Municípios de São Paulo e respectivos fóruns
}
\keyword{datasets}
