% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numero.R
\name{numero}
\alias{numero}
\title{Converte valor da ação para número ou qualquer outra string.}
\usage{
numero(str)
}
\arguments{
\item{str}{vetor de strings}
}
\value{
mesmo vetor porém numérico.
}
\description{
Converte valor da ação para número ou qualquer outra string.
}
