% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_cpopg_oab.R
\name{obter_cpopg_oab}
\alias{obter_cpopg_oab}
\title{Busca os números de processos no primeiro grau com base na oab do advogado}
\usage{
obter_cpopg_oab(oab = NULL)
}
\arguments{
\item{oab}{Número da oab seguido pela uf, ex. 123456SP}
}
\value{
tibble com os números dos processos
}
\description{
Busca os números de processos no primeiro grau com base na oab do advogado
}
\examples{
\dontrun{
obter_cpopg_oab(oab="123456SP")
}
}
