% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_cpopg_par.R
\name{obter_cpopg_par}
\alias{obter_cpopg_par}
\title{Busca os números de processos no primeiro grau no parametro especificado}
\usage{
obter_cpopg_par(oab = NULL, parte = NULL)
}
\arguments{
\item{oab}{Número da oab seguido pela uf, ex. 123456SP}

\item{parte}{Nome da parte}
}
\value{
tibble com os números dos processos
}
\description{
Busca os números de processos no primeiro grau no parametro especificado
}
\examples{
\dontrun{
obter_cpopg_par(oab="123456SP")
obter_cpopg_par(parte = "José da Silva Pereira")
}
}
