% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordena_camaras.R
\name{ordenar_camaras}
\alias{ordenar_camaras}
\title{Limpa e ordena câmaras de direito}
\usage{
ordenar_camaras(x)
}
\arguments{
\item{x}{Vetor de câmaras}
}
\value{
Mesmo vetor convertido em fator e ordenado da primeira à
    última câmara
}
\description{
Limpa e ordena câmaras de direito
}
