% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organiza_dados_cpopg.R
\name{organizar_dados_cpopg}
\alias{organizar_dados_cpopg}
\title{Organiza metadados cpopg (apenas criminal por hora)}
\usage{
organizar_dados_cpopg(df, excluir = "")
}
\arguments{
\item{df}{data.frame lido com ler_dados_cpopg}

\item{excluir}{informar vetor de frases a serem excluídas dos assuntos.
Há uma sugestão de palavras no dataset 'excluir_assuntos_cpopg'}
}
\value{
data.frame limpo e organizado
}
\description{
Organiza metadados cpopg (apenas criminal por hora)
}
