% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pontuar_cnj.R
\name{pontuar_cnj}
\alias{pontuar_cnj}
\title{Pontuar número do CNJ}
\usage{
pontuar_cnj(numero = NULL)
}
\arguments{
\item{numero}{Número pontuado ou não}
}
\value{
Número pontuado
}
\description{
Pontuar número do CNJ
}
\examples{
pontuar_cnj("12345678920218260000")
}
