% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_cjpg.R
\name{psql_cjpg}
\alias{psql_cjpg}
\title{Applies functions baixar_cjpg from package tjsp and sends data to psql db}
\usage{
psql_cjpg(con, tbl, ...)
}
\arguments{
\item{con}{PostgreSQL db connection}

\item{tbl}{Table name. It requires that have already created it.}

\item{...}{Other arguments passed to `tjsp::baixar_cjpg`, except directory.}
}
\value{
It doesn't return anything.
}
\description{
The dbAppendTable() method assumes that the table has been created beforehand,
}
\examples{
\dontrun{
con<-dbx::dbxConnect()
DBI::dbCreateTable(con,"cjpg",
                  fields=c("processo", "classe", "assunto", "magistrado", "comarca", "foro", 
                               "vara", "disponibilizacao", "julgado"))
 psql_cjpg(con,"cjpg",livre="feminicidio")                      
}
}
