% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remover_acentos.R
\name{remover_acentos}
\alias{remover_acentos}
\title{Remove acentos}
\usage{
remover_acentos(x)
}
\arguments{
\item{x}{vetor de caracteres}
}
\value{
mesmo vetor com acentos removidos
}
\description{
Remove acentos
}
