% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remover_assinatura.R
\name{remover_assinatura}
\alias{remover_assinatura}
\title{Remove texto com assinatura eletrônica}
\usage{
remover_assinatura(x)
}
\arguments{
\item{x}{vetor de textos}
}
\value{
vetor de textos sem assinatura
}
\description{
Remove texto com assinatura eletrônica
}
