% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remover_assinatura_cjpg.R
\name{remover_assinatura_cjpg}
\alias{remover_assinatura_cjpg}
\title{Remove assinutura das sentencas}
\usage{
remover_assinatura_cjpg(x)
}
\arguments{
\item{x}{Texto}
}
\value{
Texto com assinaturas removidas
}
\description{
Remove assinutura das sentencas
}
