% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remover_vazios_cpopg.R
\name{remover_vazios_cpopg}
\alias{remover_vazios_cpopg}
\title{Remove cpopg htmls vazios, ou seja, htmls menores que 90Kb.}
\usage{
remover_vazios_cpopg(diretorio = ".")
}
\arguments{
\item{diretorio}{Default para diretorio atual.}
}
\value{
Mantidos apenas arquivos com conteúdo
}
\description{
Remove cpopg htmls vazios, ou seja, htmls menores que 90Kb.
}
