% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_dup.R
\name{rm_dup}
\alias{rm_dup}
\title{Função rm_dup}
\usage{
rm_dup(x, y)
}
\arguments{
\item{x}{data.frame}

\item{y}{coluna a serem idenficiados as duplicacões.}
}
\description{
Esta função remove as linhas de um data.frame, quando há duplicidade em uma coluna.
}
\examples{
rm_dup(df,"docname")
}
\keyword{duplicated,}
\keyword{remove}

