% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_duplicados.R
\name{rm_duplicados}
\alias{rm_duplicados}
\title{Remove as linhas de um dataframe, quando há duplicidade na coluna especificada.}
\usage{
rm_duplicados(x, y)
}
\arguments{
\item{x}{data.frame}

\item{y}{coluna com células a serem identificadas como duplicadas.}
}
\description{
Remove as linhas de um dataframe, quando há duplicidade na coluna especificada.
}
\examples{

rm_duplicados(iris, Species)
}
\keyword{duplicated,}
\keyword{remove}
