% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializar.R
\name{serializar}
\alias{serializar}
\title{Serializa htmls baixados.}
\usage{
serializar(origem = NULL, destino = ".", prefix = NULL)
}
\arguments{
\item{origem}{diretorio onde se encontram os htmls}

\item{destino}{diretorio para onde irão os htmls}

\item{prefix}{prefixar os arquivos com um nome.}
}
\value{
Salva htmls serializados
}
\description{
Serializa htmls baixados.
}
