% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsg_meta.R
\name{tjsg_meta}
\alias{tjsg_meta}
\title{Função tjsg_meta}
\usage{
tjsg_meta(livre, classes = "", inicio = "", fim = "")
}
\description{
Esta função extrai metadados das decisões de segundo grau do TJSP
}
\examples{
tjsg_meta(livre='\\"lei maria da penha\\"',classes="",inicio="20/06/2012", fim="01/08/2012")
}
\keyword{tjsp}
