% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp2_baixar_docs_cd_processo.R
\name{tjsp2_baixar_docs_cd_processo}
\alias{tjsp2_baixar_docs_cd_processo}
\title{Baixar petições e docs a partir da tabela lida com tjsp_ler_tabela_docs}
\usage{
tjsp2_baixar_docs_cd_processo(df, cookies_path = NULL, diretorio = ".")
}
\arguments{
\item{df}{dataframe com colunas processo, id_doc e url_doc}

\item{cookies_path}{informar o caminho para o cookie.}

\item{diretorio}{Diretório}
}
\value{
pdfs
}
\description{
Baixar petições e docs a partir da tabela lida com tjsp_ler_tabela_docs
}
