% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp2_baixar_docs_cd_processo.R
\name{tjsp2_baixar_docs_cd_processo1}
\alias{tjsp2_baixar_docs_cd_processo1}
\title{Função interna do tjsp_baixar_cpopg_docs}
\usage{
tjsp2_baixar_docs_cd_processo1(
  cd_processo = NULL,
  id_doc = NULL,
  pagina_inicial = NULL,
  pagina_final = NULL,
  urls = NULL,
  cookies_path = NULL,
  diretorio = NULL
)
}
\arguments{
\item{cd_processo}{Número do processo}

\item{id_doc}{Id do documento}

\item{pagina_inicial}{pagina inicial}

\item{pagina_final}{pagina final}

\item{urls}{Urls dos documentos}

\item{cookies_path}{NULL para pegar da requisição anterior ou
informar caminho para o cookie.}

\item{diretorio}{Diretório}
}
\value{
pdfs
}
\description{
Função interna do tjsp_baixar_cpopg_docs
}
