% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimador.R
\name{tjsp_api_extrair}
\alias{tjsp_api_extrair}
\title{Extrai dataframe da api do TJSP, função auxiliar.}
\usage{
tjsp_api_extrair(processos)
}
\arguments{
\item{processos}{Vetor de processos}
}
\value{
data.frame
}
\description{
Extrai dataframe da api do TJSP, função auxiliar.
}
