% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_api_ler.R
\name{tjsp_api_ler}
\alias{tjsp_api_ler}
\title{Lê dados da api}
\usage{
tjsp_api_ler(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório}
}
\value{
Tibble
}
\description{
Lê dados da api
}
