% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimador.R
\name{tjsp_api_quantidade}
\alias{tjsp_api_quantidade}
\title{Estima quantidade de processos com base na api do tjsp}
\usage{
tjsp_api_quantidade(inicio = NULL, fim = NULL, ano = NULL, distribuidor = NULL)
}
\arguments{
\item{inicio}{Valor inicial}

\item{fim}{Valor final}

\item{ano}{Ano}

\item{distribuidor}{código do distribuidor}
}
\value{
valor
}
\description{
Estima quantidade de processos com base na api do tjsp
}
