% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_acordaos_cjsg.R
\name{tjsp_baixar_acordaos_cjsg}
\alias{tjsp_baixar_acordaos_cjsg}
\title{Baixar acórdãos do TJSP a partir do id (cdacordo)}
\usage{
tjsp_baixar_acordaos_cjsg(cdacordao, diretorio = ".")
}
\arguments{
\item{cdacordao}{Id do acórdão}

\item{diretorio}{Diretório}
}
\value{
pdf
}
\description{
Baixar acórdãos do TJSP a partir do id (cdacordo)
}
