% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_acordaos_cposg.R
\name{tjsp_baixar_acordaos_cposg}
\alias{tjsp_baixar_acordaos_cposg}
\alias{baixar_acordaos}
\title{Baixa acórdãos do Tribunal de Justiça de São Paulo a partir do andamento}
\usage{
tjsp_baixar_acordaos_cposg(processos = NULL, diretorio = ".")

baixar_acordaos(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{vetor com o número do processo com ou sem os separadores}

\item{diretorio}{Diretório. Default para o corrente.}
}
\value{
tabela com metadados do pdf e pdf do acórdão baixado com data
   do acordao e número do processo.
}
\description{
Baixa acórdãos do Tribunal de Justiça de São Paulo a partir do andamento
}
