% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cjpg.R
\name{tjsp_baixar_cjpg}
\alias{tjsp_baixar_cjpg}
\alias{baixar_cjpg}
\title{Baixa julgados de primeiro grau de São Paulo}
\usage{
tjsp_baixar_cjpg(
  livre = "",
  aspas = FALSE,
  processo = "",
  foro = "",
  vara = "",
  classe = "",
  assunto = "",
  magistrado = "",
  inicio = "",
  fim = "",
  diretorio = ".",
  paginas = NULL
)

baixar_cjpg(
  livre = "",
  aspas = FALSE,
  processo = "",
  foro = "",
  vara = "",
  classe = "",
  assunto = "",
  magistrado = "",
  inicio = "",
  fim = "",
  diretorio = ".",
  paginas = NULL
)
}
\arguments{
\item{livre}{Busca livre. Usar os mesmos critérios da página.}

\item{aspas}{Lógico. Coloque TRUE se quiser colocar aspas nas busca livre.}

\item{processo}{Número do processo unificado. Pode colocar com pontos ou sem pontos.}

\item{foro}{Número do fórum}

\item{vara}{Número da vara}

\item{classe}{Número da classe processual}

\item{assunto}{Número do assunto}

\item{magistrado}{Código do magistrado}

\item{inicio}{Data inicial no formato "dd/mm/aaa"}

\item{fim}{Data final no formato "dd/mm/aaaa" O lapso entre essas duas datas não deve passar de um ano.}

\item{diretorio}{Diretório onde serão armazenados os htmls.}

\item{paginas}{Páginas a serem baixadas}
}
\value{
Baixa os htmls com metadados das decisãoes de primeiro grau.
}
\description{
Baixa julgados de primeiro grau de São Paulo
}
\examples{
\dontrun{
tjsp_baixar_cjpg("homicídio simples")
}

}
