% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cjsg.R
\name{tjsp_baixar_cjsg1}
\alias{tjsp_baixar_cjsg1}
\title{Qualquer data}
\usage{
tjsp_baixar_cjsg1(
  livre = "",
  ementa = "",
  processo = "",
  classe = "",
  assunto = "",
  orgao_julgador = "",
  inicio = "",
  fim = "",
  inicio_pb = "",
  fim_pb = "",
  comarca = "",
  sg = "T",
  cr = "",
  tipo = "A",
  n = NULL,
  diretorio = ".",
  aspas = FALSE
)
}
\arguments{
\item{livre}{palavra ou texto a ser buscado nas ementas e nos acórdãos}

\item{ementa}{palavra ou texto a ser buscado apenas nas ementas}

\item{processo}{Número do processo}

\item{classe}{Código da classe processual}

\item{assunto}{Código do assunto}

\item{orgao_julgador}{Código do órgão julgador}

\item{inicio}{data inicial julgamento}

\item{fim}{Data final julgamento}

\item{inicio_pb}{data inicial registro/publicação}

\item{fim_pb}{data final registr/publicacao}

\item{comarca}{Vetor de códigos das comarcas}

\item{sg}{"T" para origem segundo grau}

\item{cr}{"R" para origem colégio recursal}

\item{tipo}{"A" Para acórdãos, "D" para decisões monocráticas}

\item{n}{Número de páginas}

\item{diretorio}{Diretório onde serão armazenadas as páginas html}

\item{aspas}{lógico. Colocar a expressão entre aspas?}
}
\value{
htmls
}
\description{
Qualquer data
}
