% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cpopg_cd_processo.R
\name{tjsp_baixar_cpopg_cd_processo}
\alias{tjsp_baixar_cpopg_cd_processo}
\title{Baixa dados processuais com base no c贸digo do processo}
\usage{
tjsp_baixar_cpopg_cd_processo(cd_processo, diretorio = ".")
}
\arguments{
\item{cd_processo}{C贸digo do processo. Possivelmente, a coluna
cd_doc do cjpg obtida com tjsp_ler_cjpg pode ser usada.}

\item{diretorio}{Diret贸rio onde armazenar os arquivos.}
}
\value{
htmls
}
\description{
Baixa dados processuais com base no c贸digo do processo
}
