% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cpopg_par.R
\name{tjsp_baixar_cpopg_par}
\alias{tjsp_baixar_cpopg_par}
\title{Busca os números de processos no primeiro grau no parametro especificado}
\usage{
tjsp_baixar_cpopg_par(
  consulta = NULL,
  parametro = NULL,
  distribuidor = "-1",
  diretorio = "."
)
}
\arguments{
\item{consulta}{Valor da consulta}

\item{parametro}{Uma dessas: "NUMOAB", "NMPARTE", "DOCPARTE"}

\item{distribuidor}{Informar código do distribuidor.
Você pode obtê-los com `tjsp_varas()`}

\item{diretorio}{Informar diretorio onde armazenar os htmls}
}
\value{
htmls
}
\description{
Busca os números de processos no primeiro grau no parametro especificado
}
\examples{
\dontrun{
tjsp_baixar_cpopg_par(consulta="123456SP", parametro = "NUMOAB")
obter_baixar_cpopg_par(consulta = "José da Silva Pereira", parametro = "NMPARTE")
}
}
