% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cposg.R
\name{tjsp_baixar_cposg}
\alias{tjsp_baixar_cposg}
\title{Baixar processos de segunda instância do tjsp}
\usage{
tjsp_baixar_cposg(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{processos de segunda instância}

\item{diretorio}{diretório}
}
\value{
html com dados processuais
}
\description{
Baixar processos de segunda instância do tjsp
}
\details{
Alguns processos podem baixar também os incidentes ou
    dependências ou desdobramentos. O cd_processo serve
    para distingui-los. Baixe os seguintes processos para ver:
    c("10006289720178260673",
    "00000066120098260270",
    "15014338820228260617")
}
