% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_doc_movimentacao.R
\name{tjsp_baixar_doc_movimentacao}
\alias{tjsp_baixar_doc_movimentacao}
\title{Baixa documentos da movimentação a partir da url}
\usage{
tjsp_baixar_doc_movimentacao(urls, diretorio = ".")
}
\arguments{
\item{urls}{Vetor de urls}

\item{diretorio}{}
}
\value{
PDF
}
\description{
Baixa documentos da movimentação a partir da url
}
