% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cpopg_docs.R
\name{tjsp_baixar_docs}
\alias{tjsp_baixar_docs}
\title{Função interna do tjsp_baixar_cpopg_docs}
\usage{
tjsp_baixar_docs(
  processos = NULL,
  id_doc = NULL,
  pagina = NULL,
  urls = NULL,
  diretorio = NULL
)
}
\arguments{
\item{processos}{Número do processo}

\item{id_doc}{Id do documento}

\item{pagina}{Pagina}

\item{urls}{Urls dos documentos}

\item{diretorio}{Diretório}
}
\value{
pdfs
}
\description{
Função interna do tjsp_baixar_cpopg_docs
}
