% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cpopg_docs_cd_processo.R
\name{tjsp_baixar_docs_cd_processo}
\alias{tjsp_baixar_docs_cd_processo}
\title{Baixar petições e docs a partir da tabela lida com tjsp_ler_tabela_docs}
\usage{
tjsp_baixar_docs_cd_processo(df, diretorio = ".")
}
\arguments{
\item{df}{dataframe com colunas processo, id_doc e url_doc}

\item{diretorio}{Diretório}
}
\value{
pdfs
}
\description{
Baixar petições e docs a partir da tabela lida com tjsp_ler_tabela_docs
}
