% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cpopg_docs_cd_processo.R
\name{tjsp_baixar_docs_cd_processo1}
\alias{tjsp_baixar_docs_cd_processo1}
\title{Função interna do tjsp_baixar_cpopg_docs}
\usage{
tjsp_baixar_docs_cd_processo1(
  cd_processo = NULL,
  id_doc = NULL,
  pagina_inicial = NULL,
  urls = NULL,
  diretorio = NULL
)
}
\arguments{
\item{cd_processo}{Número do processo}

\item{id_doc}{Id do documento}

\item{pagina_inicial}{Pagina}

\item{urls}{Urls dos documentos}

\item{diretorio}{Diretório}
}
\value{
pdfs
}
\description{
Função interna do tjsp_baixar_cpopg_docs
}
