% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_eproc_jurisprudencia.R
\name{tjsp_baixar_eproc1g_jurisprudencia}
\alias{tjsp_baixar_eproc1g_jurisprudencia}
\title{Baixa jurisprudência do sistema eproc de primeiro grau do TJSP}
\usage{
tjsp_baixar_eproc1g_jurisprudencia(
  pesquisa = "",
  campo = c("I", "E"),
  agrupar = TRUE,
  precedente = FALSE,
  colegio_recursal = FALSE,
  tipo_documento = character(0),
  processo = "",
  classe = character(0),
  relator = character(0),
  orgao = character(0),
  inicio = "",
  fim = "",
  inicio_pb = "",
  fim_pb = "",
  n = NULL,
  diretorio = "."
)
}
\arguments{
\item{pesquisa}{Texto livre para pesquisa}

\item{campo}{"I" para inteiro teor (padrão) ou "E" para somente ementa}

\item{agrupar}{Lógico. Agrupar resultados? Padrão TRUE.}

\item{precedente}{Lógico. Filtrar somente precedentes relevantes? Padrão FALSE.}

\item{colegio_recursal}{Lógico. Incluir somente Colégio Recursal? Padrão FALSE.}

\item{tipo_documento}{Vetor de caracteres com tipos de documento.}

\item{processo}{Número do processo.}

\item{classe}{Vetor de caracteres com nomes das classes processuais.}

\item{relator}{Vetor de caracteres com nomes dos relatores.}

\item{orgao}{Vetor de caracteres com nomes dos órgãos julgadores.}

\item{inicio}{Data inicial de decisão no formato "dd/mm/aaaa".}

\item{fim}{Data final de decisão no formato "dd/mm/aaaa".}

\item{inicio_pb}{Data inicial de publicação no formato "dd/mm/aaaa".}

\item{fim_pb}{Data final de publicação no formato "dd/mm/aaaa".}

\item{n}{Número máximo de páginas. Se NULL, baixa todas.}

\item{diretorio}{Diretório para salvar os arquivos HTML. Padrão é o diretório atual.}
}
\value{
Salva arquivos HTML com os resultados no diretório especificado.
}
\description{
Baixa jurisprudência do sistema eproc de primeiro grau do TJSP
}
\examples{
\dontrun{
tjsp_baixar_eproc1g_jurisprudencia(pesquisa = "homicídio", n = 1, diretorio = tempdir())
}
}
