% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_eproc_jurisprudencia.R
\name{tjsp_baixar_eproc_inteiro_teor}
\alias{tjsp_baixar_eproc_inteiro_teor}
\title{Baixa o inteiro teor de documentos do eproc}
\usage{
tjsp_baixar_eproc_inteiro_teor(
  id_jurisprudencia,
  grau = c("2g", "1g"),
  diretorio = "."
)
}
\arguments{
\item{id_jurisprudencia}{Vetor de IDs obtidos da coluna \code{id_jurisprudencia}.}

\item{grau}{"1g" ou "2g".}

\item{diretorio}{Diretório onde salvar os arquivos HTML.}
}
\value{
Salva arquivos HTML no diretório especificado.
}
\description{
Usa os \code{id_jurisprudencia} obtidos por \code{tjsp_ler_eproc_jurisprudencia}
para baixar o inteiro teor dos documentos.
}
\examples{
\dontrun{
df <- tjsp_ler_eproc_jurisprudencia(diretorio = "dados")
acordaos <- dplyr::filter(df, tipo == "Acórdão")
tjsp_baixar_eproc_inteiro_teor(acordaos$id_jurisprudencia, grau = "2g", diretorio = "inteiros_teores")
}
}
