% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_eproc_jurisprudencia.R
\name{tjsp_baixar_eproc_jurisprudencia1}
\alias{tjsp_baixar_eproc_jurisprudencia1}
\title{Worker interno para download de jurisprudência eproc}
\usage{
tjsp_baixar_eproc_jurisprudencia1(
  base_url,
  pesquisa,
  campo,
  agrupar,
  precedente,
  colegio_recursal,
  tipo_documento,
  processo,
  classe,
  relator,
  orgao,
  inicio,
  fim,
  inicio_pb,
  fim_pb,
  n,
  diretorio
)
}
\arguments{
\item{base_url}{URL base do sistema eproc (1g ou 2g)}

\item{pesquisa}{Texto livre para pesquisa}

\item{campo}{"I" para inteiro teor (padrão) ou "E" para somente ementa}

\item{agrupar}{Lógico. Agrupar resultados? Padrão TRUE.}

\item{precedente}{Lógico. Filtrar somente precedentes relevantes? Padrão FALSE.}

\item{colegio_recursal}{Lógico. Incluir somente Colégio Recursal? Padrão FALSE.}

\item{tipo_documento}{Vetor de caracteres com tipos de documento.}

\item{processo}{Número do processo.}

\item{classe}{Vetor de caracteres com nomes das classes processuais.}

\item{relator}{Vetor de caracteres com nomes dos relatores.}

\item{orgao}{Vetor de caracteres com nomes dos órgãos julgadores.}

\item{inicio}{Data inicial de decisão no formato "dd/mm/aaaa".}

\item{fim}{Data final de decisão no formato "dd/mm/aaaa".}

\item{inicio_pb}{Data inicial de publicação no formato "dd/mm/aaaa".}

\item{fim_pb}{Data final de publicação no formato "dd/mm/aaaa".}

\item{n}{Número máximo de páginas. Se NULL, baixa todas.}

\item{diretorio}{Diretório para salvar os arquivos HTML. Padrão é o diretório atual.}
}
\value{
Salva arquivos HTML no diretório especificado.
}
\description{
Worker interno para download de jurisprudência eproc
}
