% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_execucao_penal.R
\name{tjsp_baixar_execucao_penal_cd_processo}
\alias{tjsp_baixar_execucao_penal_cd_processo}
\title{Baixa metadados da execução penal}
\usage{
tjsp_baixar_execucao_penal_cd_processo(cdprocesso, diretorio = ".")
}
\arguments{
\item{cdprocesso}{Código do processo}

\item{diretorio}{Diretório onde armazenar}
}
\value{
html
}
\description{
Baixa metadados da execução penal
}
