% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historico_partes.R
\name{tjsp_baixar_historico_parte_cd_processo}
\alias{tjsp_baixar_historico_parte_cd_processo}
\title{Baixa metadados do hist贸rico da parte}
\usage{
tjsp_baixar_historico_parte_cd_processo(cdprocesso, diretorio = ".")
}
\arguments{
\item{cdprocesso}{C贸digo do processo}

\item{diretorio}{Diret贸rio onde armazenar}
}
\value{
html
}
\description{
Baixa metadados do hist贸rico da parte
}
