% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_inicial.R
\name{tjsp_baixar_inicial}
\alias{tjsp_baixar_inicial}
\title{Baixa petição inicial, ajuizamento (jec) ou denúncia}
\usage{
tjsp_baixar_inicial(processo, diretorio)
}
\arguments{
\item{processo}{Número do processo}

\item{diretorio}{Diretório onde armazenar a inicial}
}
\value{
PDF?
}
\description{
Baixa petição inicial, ajuizamento (jec) ou denúncia
}
