% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_intimacoes.R
\name{tjsp_baixar_intimacoes}
\alias{tjsp_baixar_intimacoes}
\title{Coleta intimações do TJSP}
\usage{
tjsp_baixar_intimacoes(
  cd_foro = "",
  cd_instancia = 1,
  cd_vara = "",
  dt_inicio = "",
  dt_fim = "",
  situacao = "ambas",
  natureza_comunicacao = "ambas",
  versao = c("xml", "csv"),
  intervalo = "trimestral",
  diretorio = "."
)
}
\arguments{
\item{cd_foro}{Código do Foro}

\item{cd_instancia}{Código da instância. Padrão: 1}

\item{cd_vara}{Código da vara.}

\item{dt_inicio}{Data do início no formato "dd/mm/aaaa".
Se não informada, usa a data de hoje.}

\item{dt_fim}{Data do fim no formato "dd/mm/aaaa".
Se não informada, uda a data de hoje.}

\item{situacao}{Opções: "cumprida", "pendente" ou "ambas". Padrão para "ambas".}

\item{natureza_comunicacao}{Opções: "intimação", "citação" ou "ambas". Padrão para "ambas".}

\item{versao}{Pode ser csv ou xml. Padrão para csv.}

\item{intervalo}{Padrão para trimestral, ou seja, máximo.}

\item{diretorio}{Padrão ".", ou seja, o atual.}
}
\value{
xml
}
\description{
Coleta intimações do TJSP
}
