% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_intimacoes_csv.R
\name{tjsp_baixar_intimacoes_csv}
\alias{tjsp_baixar_intimacoes_csv}
\title{Coleta intimações do TJSP}
\usage{
tjsp_baixar_intimacoes_csv(
  cd_foro = "",
  cd_instancia = 1,
  cd_vara = "",
  dt_inicio = "",
  dt_fim = "",
  situacao = "ambas",
  natureza_comunicacao = "ambas",
  intervalo = "trimestral",
  diretorio = "."
)
}
\arguments{
\item{cd_foro}{Código do Foro}

\item{cd_instancia}{Código da instância. Padrão: 1}

\item{cd_vara}{Código da vara.}

\item{dt_inicio}{Data do início no formato "dd/mm/aaaa"}

\item{dt_fim}{Data do fim no formato "dd/mm/aaaa"}

\item{situacao}{Opções: "cumprida", "pendente" ou "ambas". Padrão para "ambas".}

\item{natureza_comunicacao}{Opções: "intimação", "citação" ou "ambas". Padrão para "ambas".}

\item{intervalo}{Padrão para semestral. Pode ser mensal também.}

\item{diretorio}{Padrão ".", ou seja, o atual.}
}
\value{
csv
}
\description{
Coleta intimações do TJSP
}
