% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_partes_docs.R
\name{tjsp_baixar_partes_docs}
\alias{tjsp_baixar_partes_docs}
\title{Baixa informações sobre as partes}
\usage{
tjsp_baixar_partes_docs(cd_processo, instancia = "PG", diretorio = ".")
}
\arguments{
\item{cd_processo}{Vetor com os códigos dos processos}

\item{instancia}{Uma dessas c("PG","SG","CR"), primeiro grau, segundo grau, colégio}

\item{diretorio}{Diretório onde armazenar os jsons.}
}
\value{
Arquivo json
}
\description{
Baixa informações sobre as partes
}
\details{
Para usar esta função, você tem de estar autenticado com a função `tjsp_autenticar` e
      criar variáveis de ambiente para cd_usuario,
     usuario_solicitante, documento_usuario e cd_perfil.
     Você pode usar `usethis::edit_r_environ` para criá-las permanentemente.
     Estas são as variável que você deve criar: ESAJ_CD_USUARIO,
     ESAJ_USUARIO_SOLICITANTE, ESAJ_DOCUMENTO_USUARIO, ESAJ_CD_PERFIL.
     Alternativamente,  usa a função `tjsp_partes_credenciais` para criá-las somente para
     a presente sessão do R.
}
