% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_sentenca_cjpg.R
\name{tjsp_baixar_sentenca_cjpg}
\alias{tjsp_baixar_sentenca_cjpg}
\title{Baixar julgados do banco de sentenças em pdf em pdf}
\usage{
tjsp_baixar_sentenca_cjpg(cd_doc, diretorio = ".")
}
\arguments{
\item{cd_doc}{Código da sentença obtido de tjsp_ler_cjpg}

\item{diretorio}{Diretório onde armazenar os arquivos}
}
\value{
pdf
}
\description{
Baixar julgados do banco de sentenças em pdf em pdf
}
