% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_tabela_cd_processo.R
\name{tjsp_baixar_tabela_cd_processo}
\alias{tjsp_baixar_tabela_cd_processo}
\title{Baixa tabela de documentos com base no cd_processo}
\usage{
tjsp_baixar_tabela_cd_processo(cd_processo = NULL, diretorio = ".")
}
\arguments{
\item{cd_processo}{Número do processo}

\item{diretorio}{Diretório onde armazenar as tabelas}
}
\value{
html
}
\description{
Baixa tabela de documentos com base no cd_processo
}
