% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_tabela_cposg_docs.R
\name{tjsp_baixar_tabela_cposg_docs}
\alias{tjsp_baixar_tabela_cposg_docs}
\title{Baixa tabelas dos documentos do processo}
\usage{
tjsp_baixar_tabela_cposg_docs(processos, diretorio = ".")
}
\arguments{
\item{processos}{Número do processo}

\item{diretorio}{Diretório}
}
\value{
html
}
\description{
Baixa tabelas dos documentos do processo
}
