% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_classificar_recurso.R
\name{tjsp_classificar_recurso}
\alias{tjsp_classificar_recurso}
\alias{classificar_recurso2}
\title{Classifica recursos}
\usage{
tjsp_classificar_recurso(x)

classificar_recurso2(x)
}
\arguments{
\item{x}{vetor de julgados}
}
\value{
decisão
}
\description{
Classifica recursos
}
\examples{
\dontrun{
x <- tjsp_classificar_recurso(x = julgado)
}

}
