% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_classificar_sentenca.R
\name{tjsp_classificar_sentenca}
\alias{tjsp_classificar_sentenca}
\alias{classificar_sentenca2}
\title{Classifica decisões de primeiro grau}
\usage{
tjsp_classificar_sentenca(x)

classificar_sentenca2(x)
}
\arguments{
\item{x}{vetor de sentenças}
}
\value{
decisão
}
\description{
Classifica decisões de primeiro grau
}
\examples{
\dontrun{
x <- tjsp_classificar_sentenca(x = sentenca)
}
}
