% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_writ.R
\name{tjsp_classificar_writ}
\alias{tjsp_classificar_writ}
\alias{classificar_writ}
\title{Classificar writs(HC, MS, revisão criminal)}
\usage{
tjsp_classificar_writ(x)

classificar_writ(x)
}
\arguments{
\item{x}{Texto da decisão}
}
\value{
classificação
}
\description{
Classificar writs(HC, MS, revisão criminal)
}
