% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_combinar_docs.R
\name{tjsp_combinar_docs}
\alias{tjsp_combinar_docs}
\title{Combina docs baixados com tjsp_baixar_docs}
\usage{
tjsp_combinar_docs(
  arquivos = NULL,
  dir_origem = ".",
  dir_destino = NULL,
  nivel = c("processo", "doc"),
  forcar = TRUE
)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{dir_origem}{Diretório de origem, se não
informar arquivos}

\item{dir_destino}{Diretório destino}

\item{nivel}{Juntar processo inteiro ou por documento?}

\item{forcar}{Padrão para TRUE. Documentos inválidos são excluídos}
}
\value{
Pdf(s) combinado(s)
}
\description{
Combina docs baixados com tjsp_baixar_docs
}
