% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_combinar_docs_cd_processo.R
\name{tjsp_combinar_docs_cd_processo}
\alias{tjsp_combinar_docs_cd_processo}
\title{Combina docs baixados com tjsp_baixar_docs_cd_processo}
\usage{
tjsp_combinar_docs_cd_processo(
  arquivos = NULL,
  dir_origem = ".",
  dir_destino = NULL,
  nivel = c("processo", "doc")
)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{dir_origem}{Diretório de origem, se não
informar arquivos}

\item{dir_destino}{Diretório destino}

\item{nivel}{Juntar processo inteiro ou por documento?}
}
\value{
único pdf
}
\description{
Combina docs baixados com tjsp_baixar_docs_cd_processo
}
