% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_corrigir_varas.R
\name{tjsp_corrigir_varas}
\alias{tjsp_corrigir_varas}
\title{Corrigir nomes das varas}
\usage{
tjsp_corrigir_varas(df, coluna)
}
\arguments{
\item{df}{data.frame com a coluna dos nomes da varas a serem corrigidas}

\item{coluna}{nome da coluna}
}
\value{
mesmo data.frame adicionado da coluna vara2
}
\description{
Corrigir nomes das varas
}
