% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_baixar_tabela_docs.R
\name{tjsp_cpopg_baixar_tabela_docs}
\alias{tjsp_cpopg_baixar_tabela_docs}
\title{Baixa tabela de documentos dos processos do TJSP}
\usage{
tjsp_cpopg_baixar_tabela_docs(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Número do processo}

\item{diretorio}{Diretório onde armazenar as tabelas}
}
\value{
html
}
\description{
Baixa tabela de documentos dos processos do TJSP
}
