% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_intimacoes_foros.R
\name{tjsp_intimacoes_foros}
\alias{tjsp_intimacoes_foros}
\title{Extrai foros e respectivos códigos da página de intimações do Esaj.}
\usage{
tjsp_intimacoes_foros()
}
\value{
tibble
}
\description{
Extrai foros e respectivos códigos da página de intimações do Esaj.
}
