% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_intimacoes_varas.R
\name{tjsp_intimacoes_varas}
\alias{tjsp_intimacoes_varas}
\title{Obtêm códigos das varas a partir do código do foro.}
\usage{
tjsp_intimacoes_varas(codigo_foro)
}
\arguments{
\item{codigo_foro}{Código do foro obtido com tjsp_intimacoes_foros}
}
\value{
tibble
}
\description{
Obtêm códigos das varas a partir do código do foro.
}
\details{
Esta função requer que você tenha atribuído o código de usuário
   à variável de ambiente  ESAJ_CD_USUARIO
}
