% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_acordaos_cjsg.R
\name{tjsp_ler_acordaos_cjsg}
\alias{tjsp_ler_acordaos_cjsg}
\title{Ler acórdãos baixados com tjsp_baixar_acordao_cjsg}
\usage{
tjsp_ler_acordaos_cjsg(
  arquivos = NULL,
  diretorio = ".",
  remover_assinatura = TRUE,
  combinar = TRUE
)
}
\arguments{
\item{arquivos}{Arquivos dos pdfs}

\item{diretorio}{Informar se não informar arquivos}

\item{remover_assinatura}{Remover assinatura digital?}

\item{combinar}{Combinar as páginas em uma}
}
\value{
tibble
}
\description{
Ler acórdãos baixados com tjsp_baixar_acordao_cjsg
}
