% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_cjpg2.R
\name{tjsp_ler_cjpg2}
\alias{tjsp_ler_cjpg2}
\alias{ler_cjpg2}
\title{Lê decisões de primeira instância do TJSP baixadas por meio da função baixar_cjpg.}
\usage{
tjsp_ler_cjpg2(arquivos = NULL, diretorio = ".")

ler_cjpg2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se NULL, informar diretorio}

\item{diretorio}{objeto ou diretório onde se encontram as páginas.}
}
\value{
uma tibble com nove colunas: processo, classe, assunto, magistrado,comarca, foro, vara,
    disponibilizacao e julgado (texto da decisão).
}
\description{
Lê decisões de primeira instância do TJSP baixadas por meio da função baixar_cjpg.
}
\details{
Esta função é experimental. Eventualmente, substituirá a função ler_cjpg
}
\examples{
\dontrun{
tjsp_ler_cjpg2()
}
}
