% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_cpopg_par.R
\name{tjsp_ler_cpopg_par}
\alias{tjsp_ler_cpopg_par}
\title{Lê busca por paramêtro diferente do processo}
\usage{
tjsp_ler_cpopg_par(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Alternativa aos arquivos}
}
\value{
tibble
}
\description{
Lê busca por paramêtro diferente do processo
}
