% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_cpopg_partes_docs.R
\name{tjsp_ler_cpopg_partes_docs}
\alias{tjsp_ler_cpopg_partes_docs}
\title{Lê dados das partes baixados com tjsp_baixar_cpopg_partes_docs}
\usage{
tjsp_ler_cpopg_partes_docs(
  arquivos = NULL,
  diretorio = ".",
  rm_listas = FALSE,
  rm_na = FALSE
)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório se não forem informados os arquivos}

\item{rm_listas}{Remover colunas listas? Geralmente estão vazias.}

\item{rm_na}{Remover colunas em todas as linhas são NA?}
}
\value{
Tibble
}
\description{
Lê dados das partes baixados com tjsp_baixar_cpopg_partes_docs
}
