% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_dados_cposg.R
\name{tjsp_ler_dados_cposg}
\alias{tjsp_ler_dados_cposg}
\alias{ler_dados_cposg}
\title{Lê metadados dos processos de segunda instância}
\usage{
tjsp_ler_dados_cposg(arquivos = NULL, diretorio = ".")

ler_dados_cposg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos. Se NULL, informar diretório.}

\item{diretorio}{Diretório onde se encontram os htmls. Informar
apenas se os arquivos não forem informados.}
}
\value{
tabela com dados do processo
}
\description{
Lê metadados dos processos de segunda instância
}
\examples{
\dontrun{
dados <- ler_dados_cposg()
}
}
