% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_dispositivo.R
\name{tjsp_ler_dispositivo}
\alias{tjsp_ler_dispositivo}
\title{Ler dispositivo do cposg}
\usage{
tjsp_ler_dispositivo(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretorio se não informar arquivos}
}
\value{
Tibble
}
\description{
Ler dispositivo do cposg
}
