% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_documentos.R
\name{tjsp_ler_documentos}
\alias{tjsp_ler_documentos}
\title{Lê documentos baixados combinados com tjsp_combinar_docs}
\usage{
tjsp_ler_documentos(
  arquivos = NULL,
  diretorio = ".",
  remover_assinatura = FALSE,
  combinar = TRUE,
  remover_cabecalho = FALSE
)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Alternativamente informar diretorio}

\item{remover_assinatura}{Remover assinatura? TRUE default}

\item{combinar}{Combinar o vetor em um documento? TRUE defalt}

\item{remover_cabecalho}{Remover o cabeçalho? TRUE default}
}
\value{
tibble
}
\description{
Lê documentos baixados combinados com tjsp_combinar_docs
}
