% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_execucao_penal.R
\name{tjsp_ler_execucao_penal_cd_processo}
\alias{tjsp_ler_execucao_penal_cd_processo}
\title{Lê os metadados da execução penal}
\usage{
tjsp_ler_execucao_penal_cd_processo(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório}
}
\value{
tibble
}
\description{
Lê os metadados da execução penal
}
