% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_execucao_penal.R
\name{tjsp_ler_execucao_penal_cd_processo1}
\alias{tjsp_ler_execucao_penal_cd_processo1}
\title{L\u00EA apenas um arquivo da execu\u00E7\u00E3o}
\usage{
tjsp_ler_execucao_penal_cd_processo1(arquivo)
}
\arguments{
\item{arquivo}{Caminho para arquivo}
}
\value{
tibble
}
\description{
L\u00EA apenas um arquivo da execu\u00E7\u00E3o
}
