% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_historico_classes.R
\name{tjsp_ler_historico_classes}
\alias{tjsp_ler_historico_classes}
\title{Extrai o histórico das classes processuais}
\usage{
tjsp_ler_historico_classes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{se não informados, informar diretório}

\item{diretorio}{objeto ou diretorio  onde se encontram os htmls}
}
\value{
tibble com alterações das classes processuais
}
\description{
Extrai o histórico das classes processuais
}
