% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_intimacoes_csv.R
\name{tjsp_ler_intimacoes}
\alias{tjsp_ler_intimacoes}
\title{Lê csv baixado com tjsp_baixar_intimacoes}
\usage{
tjsp_ler_intimacoes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Alternativamente, informar diretório.}
}
\value{
Tibble
}
\description{
Lê csv baixado com tjsp_baixar_intimacoes
}
