% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_intimacoes_xml.R
\name{tjsp_ler_intimacoes_xml}
\alias{tjsp_ler_intimacoes_xml}
\title{Ler intimações em xml}
\usage{
tjsp_ler_intimacoes_xml(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório como alternativa.}
}
\value{
tibble
}
\description{
Ler intimações em xml
}
