% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_movimentacao.R
\name{tjsp_ler_movimentacao}
\alias{tjsp_ler_movimentacao}
\title{Extrai a movimentação processual de primeira e de segunda instância}
\usage{
tjsp_ler_movimentacao(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{se não informados, informar diretório}

\item{diretorio}{objeto ou diretorio  onde se encontram os htmls}
}
\value{
tibble com a movimentação processual.
}
\description{
Extrai a movimentação processual de primeira e de segunda instância
}
\examples{
\dontrun{
andamento_cposg <- ler_movimentacao_cposg()
andamento_cpopg <- ler_movimentacao_cpopg()
}

}
