% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_partes.R
\name{tjsp_ler_partes}
\alias{tjsp_ler_partes}
\alias{ler_partes}
\title{Ler Partes do cpopg ou cposg}
\usage{
tjsp_ler_partes(arquivos = NULL, diretorio = ".")

ler_partes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório se não informar arquivos}
}
\value{
tibble
}
\description{
Ler Partes do cpopg ou cposg
}
\details{
São criadas quatros colunas : processo,
    tipo_parte, parte e representante. Esta última
    pode ser tanto o procurador quanto o representante
    legal.
}
