% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_partes_cd_processo.R
\name{tjsp_ler_partes_cd_processo}
\alias{tjsp_ler_partes_cd_processo}
\title{Ler Partes do cpopg ou cposg, quando baixados por cd_processo.}
\usage{
tjsp_ler_partes_cd_processo(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório se não informar arquivos}
}
\value{
tibble
}
\description{
Ler Partes do cpopg ou cposg, quando baixados por cd_processo.
}
\details{
São criadas cinco colunas : processo,cd_processo
    tipo_parte, parte e representante. Esta última
    pode ser tanto o procurador quanto o representante
    legal.
}
