% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_peticoes.R
\name{tjsp_ler_peticoes}
\alias{tjsp_ler_peticoes}
\title{Ler petições e outros dosc baixadas com tjsp_baixar_docs}
\usage{
tjsp_ler_peticoes(
  arquivos = NULL,
  diretorio = ".",
  remover_assinatura = TRUE,
  combinar = TRUE
)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório}

\item{remover_assinatura}{Remover assinatura, dafault para TRUE}

\item{combinar}{Juntar as páginas numa única string. Default para TRUE}
}
\value{
tibble
}
\description{
Ler petições e outros dosc baixadas com tjsp_baixar_docs
}
