% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_peticoes_diversas.R
\name{tjsp_ler_peticoes_diversas}
\alias{tjsp_ler_peticoes_diversas}
\title{Extrai tabela de petições diversas da busca cpopg}
\usage{
tjsp_ler_peticoes_diversas(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{se não informados, informar diretório}

\item{diretorio}{objeto ou diretorio  onde se encontram os htmls}
}
\value{
tibble com nomes das petições e datas
}
\description{
Extrai tabela de petições diversas da busca cpopg
}
