% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_sentencas_cjpg.R
\name{tjsp_ler_sentencas_cjpg}
\alias{tjsp_ler_sentencas_cjpg}
\title{Lê pdfs das sentenças, remove assinatura e eventualmente combina}
\usage{
tjsp_ler_sentencas_cjpg(
  arquivos = NULL,
  diretorio = ".",
  remover_assinatura = FALSE,
  combinar = TRUE
)
}
\arguments{
\item{arquivos}{Arquivos onde se encontram os pdfs}

\item{diretorio}{Diretório onde se encontram os pdffs
(alternativa a arquivos)}

\item{remover_assinatura}{Padrão para FALSE}

\item{combinar}{Default para TRUE. Juntar todas as
página num único texto}
}
\value{
Tibble
}
\description{
Lê pdfs das sentenças, remove assinatura e eventualmente combina
}
