% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_tabela_docs_cd_processo.R
\name{tjsp_ler_tabela_docs_cd_processo}
\alias{tjsp_ler_tabela_docs_cd_processo}
\title{Ler tabelas baixadas com tjsp_baixar_tabela_docs}
\usage{
tjsp_ler_tabela_docs_cd_processo(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não informar diretório}
}
\value{
tibble
}
\description{
Ler tabelas baixadas com tjsp_baixar_tabela_docs
}
