% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_tabela_incidentes.R
\name{tjsp_ler_tabela_incidentes}
\alias{tjsp_ler_tabela_incidentes}
\title{Extrai a tabela de incidentes}
\usage{
tjsp_ler_tabela_incidentes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{se não informados, informar diretório}

\item{diretorio}{objeto ou diretorio  onde se encontram os htmls}
}
\value{
Tibble com urls dos incidentes
}
\description{
Extrai a tabela de incidentes
}
