% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_url_historico_partes.R
\name{tjsp_ler_url_historico_partes}
\alias{tjsp_ler_url_historico_partes}
\title{Extrai o histórico o link para histórico das partes}
\usage{
tjsp_ler_url_historico_partes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{se não informados, informar diretório}

\item{diretorio}{objeto ou diretorio  onde se encontram os htmls}
}
\value{
tibble com urls das partes
}
\description{
Extrai o histórico o link para histórico das partes
}
