% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_obter_cpopg_par.R
\name{tjsp_obter_cpopg_par}
\alias{tjsp_obter_cpopg_par}
\title{Busca os números de processos no primeiro grau no parametro especificado}
\usage{
tjsp_obter_cpopg_par(consulta = NULL, parametro = NULL)
}
\arguments{
\item{consulta}{Valor da consulta}

\item{parametro}{Uma dessas: "NUMOAB", "NMPARTE", "DOCPARTE"}
}
\value{
tibble com os números dos processos
}
\description{
Busca os números de processos no primeiro grau no parametro especificado
}
\examples{
\dontrun{
obter_cpopg_par(consulta="123456SP", parametro = "NUMOAB")
obter_cpopg_par(consulta = "José da Silva Pereira", parametro = "NMPARTE")
}
}
