% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_obter_cposg_oab.R
\name{tjsp_obter_cposg_oab}
\alias{tjsp_obter_cposg_oab}
\title{Obter processos de segunda instância pela oab}
\usage{
tjsp_obter_cposg_oab(oab = NULL)
}
\arguments{
\item{oab}{Número da OAB}
}
\value{
Tibble
}
\description{
Obter processos de segunda instância pela oab
}
