% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_obter_cposg_por_adv.R
\name{tjsp_obter_cposg_por_adv}
\alias{tjsp_obter_cposg_por_adv}
\title{Obter lista de processos do advogado por nome do adv}
\usage{
tjsp_obter_cposg_por_adv(nome = NULL)
}
\arguments{
\item{nome}{Nome do advogado}
}
\value{
Tibble
}
\description{
Obter lista de processos do advogado por nome do adv
}
\details{
Essa funçáo não é vetorizada, use map_dfr
}
