% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_obter_valor_max.R
\name{tjsp_obter_valor_max}
\alias{tjsp_obter_valor_max}
\title{Obtêm valor máximo do julgado}
\usage{
tjsp_obter_valor_max(julgado)
}
\arguments{
\item{julgado}{Julgado ou trecho do julgao de interesse}
}
\value{
Valor
}
\description{
Obtêm valor máximo do julgado
}
