% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_parsear_distribuicao.R
\name{tjsp_parsear_distribuicao}
\alias{tjsp_parsear_distribuicao}
\title{Extrai data e horário da distribuicao}
\usage{
tjsp_parsear_distribuicao(x)
}
\arguments{
\item{x}{Coluna distribuicao}
}
\value{
data horário no formato yyymmddHHMMSS (Posixct)
}
\description{
Extrai data e horário da distribuicao
}
