% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_partes_docs.R
\name{tjsp_partes_credenciais}
\alias{tjsp_partes_credenciais}
\title{Cria variáveis de embiente temporárias para o usuário esaj.}
\usage{
tjsp_partes_credenciais(
  cd_usuario = NULL,
  usuario_solicitante = NULL,
  documento_usuario = NULL,
  cd_perfil = NULL
)
}
\arguments{
\item{cd_usuario}{Código do usuário. Aceita envvar ESAJ_CD_USUARIO.}

\item{usuario_solicitante}{Código do solicitante. Aceita envvar ESAJ_USUARIO_SOLICITANTE.}

\item{documento_usuario}{CPF ou CNPJ do usuário. Aceita envvar ESAJ_DOCUMENTO_USUARIO.}

\item{cd_perfil}{Perfil do usuário. Aceita envvar ESAJ_CD_PERFIL}
}
\value{
Arquivo json
}
\description{
Cria variáveis de embiente temporárias para o usuário esaj.
}
