% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_quantidade.R
\name{tjsp_quantidade}
\alias{tjsp_quantidade}
\title{Calcula quantidade de processos distribuídos por fórum}
\usage{
tjsp_quantidade(
  inicio = NULL,
  fim = NULL,
  ano = NULL,
  distribuidor = NULL,
  diretorio = "."
)
}
\arguments{
\item{inicio}{Número inicial da sequência. Depende muito do
distribuidor e da classe processual.}

\item{fim}{Inteiro grande o suficiente}

\item{ano}{Ano}

\item{distribuidor}{Verifique os códigos em codigos2}

\item{diretorio}{Um diretório vazio para downloads provisórios}
}
\value{
vetor
}
\description{
Calcula quantidade de processos distribuídos por fórum
}
