% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_votacao.R
\name{tjsp_votacao}
\alias{tjsp_votacao}
\alias{votacao}
\title{Cria vetor informando se o voto por por unanimidade ou por maioria}
\usage{
tjsp_votacao(x = NULL)

votacao(x = NULL)
}
\arguments{
\item{x}{Vetor de dispositivos}
}
\value{
Vetor
}
\description{
Cria vetor informando se o voto por por unanimidade ou por maioria
}
