% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_cjsg.R
\name{verificar_datas}
\alias{verificar_datas}
\title{Verifica se está tudo ok com as datas.}
\usage{
verificar_datas(inicio, fim, inicio_pb, fim_pb)
}
\arguments{
\item{inicio}{Data inicial de julgamento}

\item{fim}{Data final de julgamento}

\item{inicio_pb}{Data inicial de publicação}

\item{fim_pb}{Data final de publicação}
}
\value{
Retorna TRUE se tem algo errado.
}
\description{
Verifica se está tudo ok com as datas.
}
