% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacao.R
\name{votacao}
\alias{votacao}
\title{Cria vetor informando se o voto por por unanimidade ou por maioria}
\usage{
votacao(x = NULL)
}
\arguments{
\item{x}{Vetor de dispositivos}
}
\value{
Vetor
}
\description{
Cria vetor informando se o voto por por unanimidade ou por maioria
}
