% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar_assuntos.R
\name{ajustar_assuntos}
\alias{ajustar_assuntos}
\title{Ajustar os assuntos com base na tabela do TJSP}
\usage{
ajustar_assuntos(df, excluir_cartas = TRUE)
}
\arguments{
\item{df}{tibble obtido a partir da função organiza_dados_cpopg}

\item{excluir_cartas}{default para TRUE. Excluir cartas de ordem e pracatórias.
Se você optar por FALSE, verifique a compatibilidade entre o assunto e a area.}
}
\value{
mesmo df com assuntos ajustados e áreas e  subáreas acrescidas.
}
\description{
Ajustar os assuntos com base na tabela do TJSP
}
