% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autenticar.R
\name{autenticar}
\alias{autenticar}
\title{Autenticar no tjsp}
\usage{
autenticar()
}
\value{
Estabelece uma sessão, não é necessário salvar.
}
\description{
Autenticar no tjsp
}
\details{
Ao chamar esta função, abrirá duas janelas sucessivas para
    digitar cpf e senha.
}
