% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autenticar.R
\name{autenticar}
\alias{autenticar}
\title{Autenticar no tjsp}
\usage{
autenticar(login = NULL, password = NULL)
}
\arguments{
\item{login}{cpf}

\item{password}{senha}
}
\value{
Estabelece uma sessão, não é necessário salvar.
}
\description{
Autenticar no tjsp
}
\details{
Você pode informar as credenciais nos argumentos ou
     criar variáveis de ambiente: "LOGINADV" e "PASSWORDADV", ou
     chamar a função e aguardar o prompt para informar
     login e password
}
