% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cjpg2.R
\name{baixar_cjpg2}
\alias{baixar_cjpg2}
\title{Baixa decisões de primeiro grau do TJSP a partir da url criada na busca no próprio site.}
\usage{
baixar_cjpg2(url = "", diretorio = ".")
}
\arguments{
\item{url}{Faça a busca no próprio site: \link{https://esaj.tjsp.jus.br/cjpg/}}

\item{diretorio}{Diretório onde serão armazenados os htmls.}
}
\value{
Baixa os htmls com metadados das decisãoes de primeiro grau.
}
\description{
Baixa decisões de primeiro grau do TJSP a partir da url criada na busca no próprio site.
}
