% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cposg.R
\name{baixar_cposg}
\alias{baixar_cposg}
\alias{tjsp_baixar_cposg}
\title{Baixar processos de segunda inst창ncia do tjsp}
\usage{
baixar_cposg(processos = NULL, diretorio = ".")

tjsp_baixar_cposg(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{processos de segunda inst창ncia}

\item{diretorio}{diret처rio}
}
\value{
html com dados processuais
}
\description{
Baixar processos de segunda inst창ncia do tjsp
}
