% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_autenticar.R
\name{get_email_token1}
\alias{get_email_token1}
\title{Obt\u00EAm tokens do email}
\usage{
get_email_token1(email_provider, outlook, hora)
}
\arguments{
\item{email_provider}{"gmail" ou "outlook"}

\item{outlook}{Se outlook, informar se "business" ou
"personal"}

\item{hora}{Par\u00E2metro n\u00E3o usado, por ora.}
}
\value{
token
}
\description{
Obt\u00EAm tokens do email
}
