% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_acordaos.R
\name{ler_acordaos}
\alias{ler_acordaos}
\title{Lê pdfs dos acórdãos, remove assinatura e eventualmente combina}
\usage{
ler_acordaos(
  arquivos = NULL,
  diretorio = ".",
  remover_assinatura = FALSE,
  combinar = TRUE
)
}
\arguments{
\item{arquivos}{Arquivos onde se encontram os pdfs}

\item{diretorio}{Diretório onde se encontram os pdfs (alternativa a arquivos)}

\item{remover_assinatura}{Padrão para FALSE}

\item{combinar}{Default para TRUE juntar todas as páginas em um único texto.}
}
\value{
Tibble com data do acórdão, processo e texto
}
\description{
Lê pdfs dos acórdãos, remove assinatura e eventualmente combina
}
