% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cjsg.R
\name{ler_cjsg}
\alias{ler_cjsg}
\title{Ler metadados das decisões}
\usage{
ler_cjsg(path = ".")
}
\arguments{
\item{path}{Diretório onde se encontram os htmls.}
}
\value{
tabela com metadados jurisprudenciais.
}
\description{
Ler metadados das decisões
}
\examples{
cjsg<-ler_cjsg()

}
