% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cpopg.R
\name{ler_dados_cpopg}
\alias{ler_dados_cpopg}
\title{Lê os metadados de processos de primeira instância.}
\usage{
ler_dados_cpopg(fonte = ".", wide = FALSE)
}
\arguments{
\item{fonte}{objeto no global environment ou diretório onde se encontram os htmls.}

\item{wide}{o padrão é o formado longo, com apenas quatro colunas:
processo, digital, variavel, valor.}
}
\value{
tibble com os metadados
}
\description{
Lê os metadados de processos de primeira instância.
}
\examples{
\dontrun{
ler_dados_cpopg()
}

}
