% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cpopg.R
\name{ler_dados_cpopg}
\alias{ler_dados_cpopg}
\title{Lê os metadados de processos de primeira instância.}
\usage{
ler_dados_cpopg(diretorio = ".")
}
\arguments{
\item{diretorio}{Diretório onde se encontram os htmls.}
}
\value{
tibble com os metadados
}
\description{
Lê os metadados de processos de primeira instância.
}
\examples{
\dontrun{
ler_dados_cpopg()
}

}
