% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cpopg.R
\name{ler_dados_cpopg}
\alias{ler_dados_cpopg}
\title{Lê os metadados de processos de primeira instância.}
\usage{
ler_dados_cpopg(arquivos = NULL, diretorio = ".", wide = FALSE)
}
\arguments{
\item{arquivos}{Opcional. Informar se não informar o diretório}

\item{diretorio}{Diretório onde se encontram os htmls.}

\item{wide}{o padrão é o formado longo, com apenas quatro colunas:
processo, digital, variavel, valor.}
}
\value{
tibble com os metadados
}
\description{
Lê os metadados de processos de primeira instância.
}
\details{
Esta função será descontinuada. Provavelmente você
    quer usar a função tjsp_ler_dados_cpopg
}
\examples{
\dontrun{
ler_dados_cpopg()
}

}
