% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_decisoes.R
\name{ler_decisoes}
\alias{ler_decisoes}
\title{Ler decisoes}
\usage{
ler_decisoes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Caminho para os htmls baixados. Se NULL,
usar diretório}

\item{diretorio}{Diretório onde se encontram os htmls baixados}
}
\value{
tibble com as os numéros dos processos e respectivas decisões
}
\description{
Ler decisoes
}
\examples{
\dontrun{
ler_decisoes()
}
}
