% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_decisoes_monocraticas.R
\name{ler_decisoes_monocraticas}
\alias{ler_decisoes_monocraticas}
\title{Lê conteúdo das decisões monocráticas}
\usage{
ler_decisoes_monocraticas(diretorio = ".")
}
\arguments{
\item{diretorio}{objeto ou diretório onde se encontram os htmls.}
}
\value{
tibble com o número do processo, o texto da decisão e a data da
    decisão.
}
\description{
Lê conteúdo das decisões monocráticas
}
\details{
Em alguns processos, as decisões monocráticas estão etiquetadas como
    despachos. Complemente a busca com a função ler_despachos.
}
\examples{
\dontrun{
ler_decisoes_monocraticas()
}

}
