% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_entrada.R
\name{ler_entrada}
\alias{ler_entrada}
\title{Data da entrada do processo}
\usage{
ler_entrada(diretorio = ".")
}
\arguments{
\item{diretorio}{objeto ou diretório onde se encontram os htmls dos processos}
}
\value{
tibble com o número dos processos e respectivas decisões
}
\description{
Data da entrada do processo
}
\examples{
\dontrun{
entrada <- ler_entrada_cposg()
entrada <- ler_entrada_cpopg()
}

}
