% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao_cposg.R
\name{ler_movimentacao_cposg}
\alias{ler_movimentacao_cposg}
\title{Extrai a movimentação processual de primeira e de segunda instância}
\usage{
ler_movimentacao_cposg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{se não informados, informar diretório}

\item{diretorio}{objeto ou diretorio  onde se encontram os htmls}
}
\value{
tibble com a movimentação processual.
}
\description{
Extrai a movimentação processual de primeira e de segunda instância
}
\examples{
\dontrun{
andamento_cposg <- ler_movimentacao_cposg()
andamento_cpopg <- ler_movimentacao_cpopg()
}

}
