% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_partes.R
\name{ler_partes}
\alias{ler_partes}
\title{Aplica parser para extrair informações sobre as partes do
    processo.}
\usage{
ler_partes(diretorio = ".")
}
\arguments{
\item{diretorio}{Diretório onde se encontram os htmls baixados.}
}
\value{
tabela com informações das partes
}
\description{
Aplica parser para extrair informações sobre as partes do
    processo.
}
\examples{
\dontrun{
partes<-ler_partes(path=".")
}
}
