% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_partes.R
\name{ler_partes}
\alias{ler_partes}
\title{Aplica parser para extrair informa\u00e7\u00f5es sobre as partes do
    processo.}
\usage{
ler_partes(diretorio = ".")
}
\arguments{
\item{diretorio}{diret\u00f3rio onde se encontram os htmls baixados.}
}
\value{
tabela com informa\u00e7\u00f5es das partes.
}
\description{
Aplica parser para extrair informa\u00e7\u00f5es sobre as partes do
    processo.
}
