% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_dados_cpopg.R
\name{organizar_dados_cpopg}
\alias{organizar_dados_cpopg}
\title{Organiza metadados cpopg (apenas criminal por ora)}
\usage{
organizar_dados_cpopg(df, excluir_assunto = "", excluir_classe = "")
}
\arguments{
\item{df}{data.frame lido com ler_dados_cpopg}

\item{excluir_assunto}{informar vetor de frases a serem excluídas dos assuntos.
Há uma sugestão de palavras no dataset 'excluir_assuntos_cpopg'}

\item{excluir_classe}{informar vetor de frases a serem excluídas das classes.}
}
\value{
data.frame limpo e organizado
}
\description{
Organiza metadados cpopg (apenas criminal por ora)
}
