% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteiroTJSP.R
\name{testar_processo}
\alias{testar_processo}
\title{Função inteiroTJSP}
\usage{
testar_processo(url)
}
\arguments{
\item{url}{a vector of url strings to access the decisions.}
}
\value{
pdf of the decision downloaded to the working directory.
}
\description{
This function predicts captcha letters
}
\keyword{court}
\keyword{decisions}
\keyword{tjsp}
