% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjpg.R
\name{tjpg}
\alias{tjpg}
\title{Função tjpg}
\usage{
tjpg(url)
}
\arguments{
\item{url}{Faça primeiramente a busca na página do TJSP, depois copie e cole a url aqui.}
}
\value{
Um data.frame com as informações do julgado
}
\description{
Esta função extraí dados de busca livre por julgados de primeira instância no TJSP
}
\examples{
tjsg(url)
}
\keyword{instância}
\keyword{primeira}
\keyword{tjsp}
