% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjrs.R
\name{tjrs_jurisprudencia}
\alias{tjrs_jurisprudencia}
\title{Buscar jurisprudência no Tribunal de Justiça do RS}
\usage{
tjrs_jurisprudencia(
  julgamento_inicial = "",
  julgamento_final = "",
  delay = 5,
  timeout_seconds = 60,
  proxy_string = NULL
)
}
\arguments{
\item{julgamento_inicial}{Data de julgamento inicial das decisões a serem buscadas (no formato "DD/MM/YYYY"). O padrão é "" (vazio), indicando que não há restrição de data inicial.}

\item{julgamento_final}{Data de julgamento final das decisões a serem buscadas (no formato "DD/MM/YYYY"). O padrão é "" (vazio), indicando que não há restrição de data final.}

\item{delay}{Tempo de espera em segundos entre as requisições. O padrão é 5 segundos.}

\item{timeout_seconds}{Tempo máximo em segundos para esperar por uma resposta da requisição. O padrão é 60 segundos.}

\item{proxy_string}{String de conexão do proxy opcional (formato: "http://user:pass@host:port" ou "http://host:port"). Se fornecida, será usada para as requisições. O padrão é NULL (sem proxy).}
}
\value{
Uma string JSON contendo um array com as informações sobre as decisões encontradas (originalmente no campo 'docs' da resposta da API).
}
\description{
Esta função permite buscar jurisprudência no Portal de Jurisprudência do Tribunal de Justiça do Estado do Rio Grande do Sul (TJRS) com base no período de julgamento.
}
\examples{
\dontrun{
# Buscar jurisprudência tjrs
tjrs <- tjrs_jurisprudencia(julgamento_inicial = "01/01/2023", julgamento_final = "31/03/2023")
tjrs |> head(5)

# Caso ele não encontre nada, mostrará e um aviso e retornará um valor NULL
tjrs_jurisprudencia(julgamento_inicial = "01/01/2023", julgamento_final = "31/02/2023")
}
}
