% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp.R
\docType{package}
\name{tjsp-package}
\alias{tjsp}
\alias{tjsp-package}
\title{\code{tjsp} package}
\description{
Baixa  e organiza decisões do TJSP
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjesusfilho/tjsp}
  \item Report bugs at \url{https://github.com/jjesusfilho/tjsp/issues}
}

}
\author{
\strong{Maintainer}: José de Jesus Filho \email{jose@consudata.com.br}

}
