% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_doc_completo.R
\name{tjsp_baixar_doc_completo}
\alias{tjsp_baixar_doc_completo}
\title{Baixa documento completo com base no cd_processo}
\usage{
tjsp_baixar_doc_completo(cd_processo = NULL, diretorio = ".")
}
\arguments{
\item{cd_processo}{Número do processo}

\item{diretorio}{Diretório onde armazenar as tabelas}
}
\value{
pdf
}
\description{
Baixa documento completo com base no cd_processo
}
