% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_baixar_docs.R
\name{tjsp_baixar_docs}
\alias{tjsp_baixar_docs}
\title{Baixar petições e docs a partir da tabela lida com tjsp_ler_tabela_docs}
\usage{
tjsp_baixar_docs(
  processos = NULL,
  doc_id = NULL,
  urls = NULL,
  diretorio = NULL
)
}
\arguments{
\item{processos}{Número do processo}

\item{doc_id}{Id do documento}

\item{urls}{Urls dos documentos}

\item{diretorio}{Diretório}
}
\value{
pdfs
}
\description{
Baixar petições e docs a partir da tabela lida com tjsp_ler_tabela_docs
}
