% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_gerar_cd_processo.R
\name{tjsp_gerar_cd_processo}
\alias{tjsp_gerar_cd_processo}
\title{Gera cd_processo a partir de processo}
\usage{
tjsp_gerar_cd_processo(processo)
}
\arguments{
\item{processo}{Vetor com números dos processos}
}
\value{
String
}
\description{
Gera cd_processo a partir de processo
}
