% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_dados_cpopg.R
\name{tjsp_ler_dados_cpopg}
\alias{tjsp_ler_dados_cpopg}
\title{Extrai a tabela de incidentes}
\usage{
tjsp_ler_dados_cpopg(arquivos = NULL, diretorio = ".", wide = TRUE)
}
\arguments{
\item{arquivos}{se não informados, informar diretório}

\item{diretorio}{objeto ou diretorio  onde se encontram os htmls}

\item{wide}{Colocar em formato wide? Padrão TRUE}
}
\value{
tibble
}
\description{
Extrai a tabela de incidentes
}
