% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_ler_dados_cpopg.R
\name{tjsp_ler_dados_cpopg}
\alias{tjsp_ler_dados_cpopg}
\title{Lê os metadados de processos de primeira instância.}
\usage{
tjsp_ler_dados_cpopg(arquivos = NULL, diretorio = ".", wide = FALSE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar apenas se não informou arquivos}

\item{wide}{Padrão para FALSE. Colocar TRUE se quiser manter em formato wide}
}
\value{
tibble
}
\description{
Lê os metadados de processos de primeira instância.
}
