% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_execucao_penal.R
\name{tjsp_ler_execucao_penal_cd_processo1}
\alias{tjsp_ler_execucao_penal_cd_processo1}
\title{Lêapenas um arquivo da execução}
\usage{
tjsp_ler_execucao_penal_cd_processo1(arquivo)
}
\arguments{
\item{arquivo}{Caminho para arquivo}
}
\value{
tibble
}
\description{
Lêapenas um arquivo da execução
}
