% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_quantidade.R
\name{tjsp_quantidade}
\alias{tjsp_quantidade}
\title{Calcula quantidade de processos distribuídos por fórum}
\usage{
tjsp_quantidade(
  inicio = 0,
  fim = NULL,
  ano = NULL,
  distribuidor = NULL,
  funcao = "baixar_cpopg",
  diretorio = "data-raw"
)
}
\arguments{
\item{inicio}{Geralmente 1}

\item{fim}{Inteiro grande o suficiente}

\item{ano}{Ano}

\item{distribuidor}{Verifique os códigos em codigos2}

\item{funcao}{Geralmente baixar_cposg}

\item{diretorio}{Um diretório vazio para downloads provisórios}
}
\value{
vetor
}
\description{
Calcula quantidade de processos distribuídos por fórum
}
