% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacao.R
\name{votacao}
\alias{votacao}
\title{Cria coluna no dataframe informando se o voto por por unanimidade}
\usage{
votacao(df, x = NULL)
}
\arguments{
\item{df}{base lida por ler_dados}

\item{x}{coluna lida com dispositivo da decisão.}
}
\value{
mesmo dataframe adicionado da coluna votacao
}
\description{
Cria coluna no dataframe informando se o voto por por unanimidade
}
